Extract contracts app into dedicated package
