+++ src/dc43/contracts_app/demo_data/contracts/customers/1.0.0.json
