+++ src/dc43/contracts_app/static/presentation.html
