+++ src/dc43/contracts_app/templates/contract_detail.html
