+++ src/dc43/demo_app/server.py
@@ -1,2691 +1,83 @@
 from __future__ import annotations
 
-"""FastAPI demo application for dc43.
-
-This application provides a small Bootstrap-powered UI to manage data
-contracts and run an example Spark pipeline that records dataset versions
-with their validation status. Contracts are stored on the local
-filesystem using :class:`~dc43_service_backends.contracts.backend.stores.FSContractStore` and dataset
-metadata lives in a JSON file.
-
-Run the UI directly with::
-
-    uvicorn dc43.demo_app.server:app --reload
-
-or start the full demo (UI + HTTP backend) with::
-
-    dc43-demo
-
-Optional dependencies needed: ``fastapi``, ``uvicorn``, ``jinja2`` and
-``pyspark``.
-"""
-
 import asyncio
-import contextlib
 import logging
-import subprocess
-import sys
-import time
-from dataclasses import dataclass, field
-from pathlib import Path
-from typing import List, Dict, Any, Tuple, Mapping, Optional, Iterable
-from uuid import uuid4
-from threading import Lock
-from textwrap import dedent
-import json
 import os
-import re
-import shutil
-import tempfile
-from datetime import datetime
-from collections import Counter
-
-import httpx
-from fastapi import FastAPI, Request, Form, HTTPException
-from fastapi.responses import HTMLResponse, RedirectResponse
-from fastapi.staticfiles import StaticFiles
-from fastapi.templating import Jinja2Templates
-from fastapi.encoders import jsonable_encoder
+from pathlib import Path
 from urllib.parse import urlencode
-from httpx import ASGITransport
 
-from dc43_service_backends.contracts.backend.stores import FSContractStore
-from dc43_service_backends.web import build_local_app
-from dc43_service_clients.contracts.client.remote import RemoteContractServiceClient
-from dc43_service_clients.data_quality.client.remote import RemoteDataQualityServiceClient
-from dc43_service_clients.governance.client.remote import RemoteGovernanceServiceClient
-from dc43.odcs import custom_properties_dict, normalise_custom_properties
-from dc43.versioning import SemVer
-from open_data_contract_standard.model import (
-    CustomProperty,
-    DataQuality,
-    Description,
-    OpenDataContractStandard,
-    SchemaObject,
-    SchemaProperty,
-    Server,
-    ServiceLevelAgreementProperty,
-    Support,
-)
-from pydantic import ValidationError
-from packaging.version import Version, InvalidVersion
+from fastapi import FastAPI, Form, HTTPException, Request
+from fastapi.responses import HTMLResponse, RedirectResponse
+from fastapi.templating import Jinja2Templates
+from jinja2 import ChoiceLoader, Environment, FileSystemLoader, select_autoescape
 
-# Optional pyspark-based helpers. Keep imports lazy-friendly so the demo UI can
-# still load when pyspark is not installed (for example when running fast unit
-# tests).
-try:  # pragma: no cover - exercised indirectly when pyspark is available
-    from dc43_integrations.spark.io import ContractVersionLocator, read_with_contract
-except ModuleNotFoundError as exc:  # pragma: no cover - safety net for CI
-    if exc.name != "pyspark":
-        raise
-    ContractVersionLocator = None  # type: ignore[assignment]
-    read_with_contract = None  # type: ignore[assignment]
+from dc43.contracts_app import server as contracts_server
 
-_SPARK_SESSION: Any | None = None
 logger = logging.getLogger(__name__)
 
-
-def _spark_session() -> Any:
-    """Return a cached local Spark session for previews."""
-
-    global _SPARK_SESSION
-    if _SPARK_SESSION is None:
-        from pyspark.sql import SparkSession  # type: ignore
-
-        _SPARK_SESSION = (
-            SparkSession.builder.master("local[1]")
-            .appName("dc43-preview")
-            .getOrCreate()
-        )
-    return _SPARK_SESSION
+DatasetRecord = contracts_server.DatasetRecord
+SCENARIOS = contracts_server.SCENARIOS
+load_records = contracts_server.load_records
+save_records = contracts_server.save_records
+queue_flash = contracts_server.queue_flash
+pop_flash = contracts_server.pop_flash
+scenario_run_rows = contracts_server.scenario_run_rows
+set_active_version = contracts_server.set_active_version
+store = contracts_server.store
+_dq_version_records = contracts_server._dq_version_records
 
 BASE_DIR = Path(__file__).resolve().parent
-SAMPLE_DIR = BASE_DIR / "demo_data"
-WORK_DIR = Path(tempfile.mkdtemp(prefix="dc43_demo_"))
-if not os.getenv("SHOW_WORK_DIR") == "false":
-    print(f"The working dir for the demo is: {WORK_DIR}")
-    import subprocess, sys
-    if sys.platform == "darwin":
-        subprocess.run(["open", WORK_DIR])
-CONTRACT_DIR = WORK_DIR / "contracts"
-DATA_DIR = WORK_DIR / "data"
-RECORDS_DIR = WORK_DIR / "records"
-DATASETS_FILE = RECORDS_DIR / "datasets.json"
-DQ_STATUS_DIR = RECORDS_DIR / "dq_state" / "status"
-
-# Copy sample data and records into a temporary working directory so the
-# application operates on absolute paths that are isolated per run.
-shutil.copytree(SAMPLE_DIR / "data", DATA_DIR)
-shutil.copytree(SAMPLE_DIR / "records", RECORDS_DIR)
-
-
-_VERSION_PATTERN = re.compile(r"\d{4}-\d{2}-\d{2}(?:T[^_]+Z)?")
-_DERIVED_SUFFIXES = {"valid", "reject"}
-_VERSION_ALIASES = {
-    "partial": "2024-04-10",
-}
-
-
-def _normalise_dataset_layout(root: Path) -> None:
-    """Ensure datasets under ``root`` follow ``dataset/version/file`` layout."""
-
-    for candidate in list(root.glob("*.json")):
-        stem = candidate.stem
-        if "_" not in stem:
-            dataset_dir = root / stem
-            if dataset_dir.exists():
-                candidate.unlink()
-            continue
-
-        parts = stem.split("_")
-        suffix: Optional[str] = None
-        if parts[-1] in _DERIVED_SUFFIXES:
-            suffix = parts.pop()
-
-        version = parts[-1]
-        version = _VERSION_ALIASES.get(version, version)
-        if not _VERSION_PATTERN.fullmatch(version):
-            continue
-
-        dataset = "_".join(parts[:-1]) or parts[-1]
-        if not dataset:
-            continue
-
-        dataset_dir_name = dataset if suffix is None else f"{dataset}__{suffix}"
-        target_dir = root / dataset_dir_name / version
-        target_dir.mkdir(parents=True, exist_ok=True)
-
-        destination = target_dir / f"{dataset}.json"
-        if destination.exists():
-            candidate.unlink()
-        else:
-            candidate.rename(destination)
-
-
-_normalise_dataset_layout(DATA_DIR)
-
-
-def _safe_fs_name(value: str) -> str:
-    """Return a filesystem-friendly representation for governance ids."""
-
-    return "".join(ch if ch.isalnum() or ch in ("_", "-", ".") else "_" for ch in value)
-
-
-def _read_json_file(path: Path) -> Optional[Dict[str, Any]]:
-    """Return decoded JSON for ``path`` or ``None`` on failure."""
-
-    try:
-        return json.loads(path.read_text())
-    except (OSError, json.JSONDecodeError):
-        return None
-
-
-def _version_sort_key(value: str) -> tuple[int, Tuple[int, int, int] | float | str, str]:
-    """Sort versions treating ISO timestamps and SemVer intelligently."""
-
-    candidate = value
-    if candidate.endswith("Z"):
-        candidate = candidate[:-1] + "+00:00"
-    try:
-        dt = datetime.fromisoformat(candidate)
-        return (0, dt.timestamp(), value)
-    except ValueError:
-        pass
-    try:
-        parsed = SemVer.parse(value)
-        return (1, (parsed.major, parsed.minor, parsed.patch), value)
-    except ValueError:
-        return (2, value, value)
-
-
-def _sort_versions(entries: Iterable[str]) -> List[str]:
-    """Return ``entries`` sorted using :func:`_version_sort_key`."""
-
-    return sorted(entries, key=_version_sort_key)
-
-
-def _dq_status_dir_for(dataset_id: str) -> Path:
-    """Return the directory that stores compatibility statuses for ``dataset_id``."""
-
-    return DQ_STATUS_DIR / _safe_fs_name(dataset_id)
-
-
-def _dq_status_path(dataset_id: str, dataset_version: str) -> Path:
-    """Return the JSON payload path for the supplied dataset/version pair."""
-
-    directory = _dq_status_dir_for(dataset_id)
-    return directory / f"{_safe_fs_name(dataset_version)}.json"
-
-
-def _dq_status_payload(dataset_id: str, dataset_version: str) -> Optional[Dict[str, Any]]:
-    """Load the compatibility payload if available."""
-
-    path = _dq_status_path(dataset_id, dataset_version)
-    if not path.exists():
-        return None
-    return _read_json_file(path)
-
-
-def _dataset_root_for(dataset_id: str, dataset_path: Optional[str] = None) -> Optional[Path]:
-    """Return the directory that should contain materialised versions."""
-
-    base: Optional[Path] = None
-    if dataset_path:
-        try:
-            path = Path(dataset_path)
-        except (TypeError, ValueError):
-            path = None
-        if path is not None:
-            if path.suffix:
-                path = path.parent / path.stem
-            if not path.is_absolute():
-                path = (Path(DATA_DIR).parent / path).resolve()
-            base = path
-    if base is None and dataset_id:
-        base = DATA_DIR / dataset_id.replace("::", "__")
-    return base
-
-
-def _version_marker_value(folder: Path) -> str:
-    """Return the canonical version value for ``folder`` if annotated."""
-
-    marker = folder / ".dc43_version"
-    if marker.exists():
-        try:
-            text = marker.read_text().strip()
-        except OSError:
-            text = ""
-        if text:
-            return text
-    return folder.name
-
-
-def _candidate_version_paths(dataset_dir: Path, version: str) -> List[Path]:
-    """Return directories that may correspond to ``version``."""
-
-    candidates: List[Path] = []
-    direct = dataset_dir / version
-    candidates.append(direct)
-    safe = dataset_dir / _safe_fs_name(version)
-    if safe != direct:
-        candidates.append(safe)
-    try:
-        for entry in dataset_dir.iterdir():
-            if not entry.is_dir():
-                continue
-            if _version_marker_value(entry) == version and entry not in candidates:
-                candidates.append(entry)
-    except FileNotFoundError:
-        return []
-    return candidates
-
-
-def _has_version_materialisation(dataset_dir: Path, version: str) -> bool:
-    """Return ``True`` if ``dataset_dir`` contains files for ``version``."""
-
-    lowered = version.lower()
-    if lowered in {"latest", "current"} or lowered.startswith("latest__"):
-        return True
-    for candidate in _candidate_version_paths(dataset_dir, version):
-        if candidate.exists():
-            return True
-    return False
-
-
-def _existing_version_dir(dataset_dir: Path, version: str) -> Optional[Path]:
-    """Return an existing directory matching ``version`` if available."""
-
-    for candidate in _candidate_version_paths(dataset_dir, version):
-        if candidate.exists():
-            return candidate
-    return None
-
-
-def _target_version_dir(dataset_dir: Path, version: str) -> Path:
-    """Return the directory path where ``version`` should be materialised."""
-
-    safe = _safe_fs_name(version)
-    if not safe:
-        safe = "version"
-    return dataset_dir / safe
-
-
-def _ensure_version_marker(path: Path, version: str) -> None:
-    """Record ``version`` inside ``path`` for lookup when sanitised."""
-
-    if not path.exists() or not path.is_dir():
-        return
-    marker = path / ".dc43_version"
-    try:
-        marker.write_text(version)
-    except OSError:
-        pass
-
-
-def _dq_status_entries(dataset_id: str) -> List[Tuple[str, str, Dict[str, Any]]]:
-    """Return (display_version, stored_version, payload) tuples."""
-
-    directory = _dq_status_dir_for(dataset_id)
-    entries: List[Tuple[str, str, Dict[str, Any]]] = []
-    if not directory.exists():
-        return entries
-    for path in directory.glob("*.json"):
-        payload = _read_json_file(path) or {}
-        display_version = str(payload.get("dataset_version") or path.stem)
-        entries.append((display_version, path.stem, payload))
-    entries.sort(key=lambda item: _version_sort_key(item[0]))
-    return entries
-
-
-def _dq_status_versions(dataset_id: str) -> List[str]:
-    """Return known dataset versions recorded by the governance stub."""
-
-    return [entry[0] for entry in _dq_status_entries(dataset_id)]
-
-
-def _link_path(target: Path, source: Path) -> None:
-    """Create a symlink (or copy fallback) from ``target`` to ``source``."""
-
-    if target.exists() or target.is_symlink():
-        try:
-            if target.is_symlink() and target.resolve() == source.resolve():
-                return
-        except OSError:
-            pass
-        if target.is_dir() and not target.is_symlink():
-            shutil.rmtree(target)
-        else:
-            target.unlink()
-
-    target.parent.mkdir(parents=True, exist_ok=True)
-    try:
-        relative = os.path.relpath(source, target.parent)
-        target.symlink_to(relative, target_is_directory=source.is_dir())
-    except OSError:
-        if source.is_dir():
-            shutil.copytree(source, target, dirs_exist_ok=True)
-        else:
-            shutil.copy2(source, target)
-
-
-def _iter_versions(dataset_dir: Path) -> list[Path]:
-    """Return sorted dataset version directories ignoring alias folders."""
-
-    versions: list[Path] = []
-    for candidate in dataset_dir.iterdir():
-        if not candidate.is_dir():
-            continue
-        name = candidate.name
-        if name == "latest" or name.startswith("latest__"):
-            continue
-        versions.append(candidate)
-    return sorted(versions)
-
-
-def refresh_dataset_aliases(dataset: str | None = None) -> None:
-    """Populate ``latest``/derived aliases for the selected dataset(s)."""
-
-    roots: list[Path]
-    if dataset:
-        base = DATA_DIR / dataset
-        roots = [base] if base.exists() else []
-    else:
-        roots = [p for p in DATA_DIR.iterdir() if p.is_dir() and "__" not in p.name]
-
-    for dataset_dir in roots:
-        versions = _iter_versions(dataset_dir)
-        if not versions:
-            continue
-        latest = versions[-1]
-        _link_path(dataset_dir / "latest", latest)
-
-        derived_dirs = sorted(DATA_DIR.glob(f"{dataset_dir.name}__*"))
-        for derived_dir in derived_dirs:
-            if not derived_dir.is_dir():
-                continue
-            suffix = derived_dir.name.split("__", 1)[1]
-            derived_versions = _iter_versions(derived_dir)
-            for version_dir in derived_versions:
-                target = dataset_dir / version_dir.name / suffix
-                _link_path(target, version_dir)
-            if derived_versions:
-                _link_path(dataset_dir / f"latest__{suffix}", derived_versions[-1])
-
-
-def set_active_version(dataset: str, version: str) -> None:
-    """Point the ``latest`` alias of ``dataset`` (and derivatives) to ``version``."""
-
-    dataset_dir = DATA_DIR / dataset
-    target = _existing_version_dir(dataset_dir, version)
-    if target is None:
-        target = _target_version_dir(dataset_dir, version)
-    if not target.exists():
-        raise FileNotFoundError(f"Unknown dataset version: {dataset} {version}")
-
-    _link_path(dataset_dir / "latest", target)
-
-    if "__" not in dataset:
-        for derived_dir in DATA_DIR.glob(f"{dataset}__*"):
-            suffix = derived_dir.name.split("__", 1)[1]
-            derived_target = _existing_version_dir(derived_dir, version)
-            if derived_target is None:
-                continue
-            _link_path(target / suffix, derived_target)
-            _link_path(dataset_dir / f"latest__{suffix}", derived_target)
-    else:
-        base, suffix = dataset.split("__", 1)
-        base_dir = DATA_DIR / base
-        version_dir = _existing_version_dir(base_dir, version)
-        if version_dir is not None and version_dir.exists():
-            _link_path(version_dir / suffix, target)
-            _link_path(base_dir / f"latest__{suffix}", target)
-
-
-def register_dataset_version(dataset: str, version: str, source: Path) -> None:
-    """Expose ``source`` under ``data/<dataset>/<version>`` via symlink."""
-
-    dataset_dir = DATA_DIR / dataset
-    dataset_dir.mkdir(parents=True, exist_ok=True)
-    target = _target_version_dir(dataset_dir, version)
-    _link_path(target, source)
-    _ensure_version_marker(target, version)
-
-
-refresh_dataset_aliases()
-try:
-    set_active_version("customers", "2024-01-01")
-    set_active_version("orders", "2024-01-01")
-    set_active_version("orders__valid", "2025-09-28")
-    set_active_version("orders__reject", "2025-09-28")
-except FileNotFoundError:
-    # Sample data may be absent during tests that override the workspace.
-    pass
-
-# Prepare contracts with absolute server paths pointing inside the working dir.
-for src in (SAMPLE_DIR / "contracts").rglob("*.json"):
-    model = OpenDataContractStandard.model_validate_json(src.read_text())
-    for srv in model.servers or []:
-        p = Path(srv.path or "")
-        if not p.is_absolute():
-            p = (WORK_DIR / p).resolve()
-        base = p.parent if p.suffix else p
-        base.mkdir(parents=True, exist_ok=True)
-        srv.path = str(p)
-    dest = CONTRACT_DIR / src.relative_to(SAMPLE_DIR / "contracts")
-    dest.parent.mkdir(parents=True, exist_ok=True)
-    dest.write_text(
-        model.model_dump_json(indent=2, by_alias=True, exclude_none=True),
-        encoding="utf-8",
-    )
-
-store = FSContractStore(str(CONTRACT_DIR))
-
-
-_STATUS_OPTIONS: List[Tuple[str, str]] = [
-    ("", "Unspecified"),
-    ("draft", "Draft"),
-    ("active", "Active"),
-    ("deprecated", "Deprecated"),
-    ("retired", "Retired"),
-    ("suspended", "Suspended"),
-]
-
-_VERSIONING_MODES: List[Tuple[str, str]] = [
-    ("", "Not specified"),
-    ("delta", "Delta (time-travel compatible)"),
-    ("snapshot", "Snapshot folders"),
-    ("append", "Append-only log"),
-]
-
-_backend_app: FastAPI | None = None
-_backend_transport: ASGITransport | None = None
-_backend_client: httpx.AsyncClient | None = None
-contract_service: RemoteContractServiceClient
-dq_service: RemoteDataQualityServiceClient
-governance_service: RemoteGovernanceServiceClient
-
-
-def _close_backend_client() -> None:
-    """Best-effort close of the shared HTTP client."""
-
-    global _backend_client
-    client = _backend_client
-    if client is None:
-        return
-    try:
-        asyncio.run(client.aclose())
-    except RuntimeError:
-        loop = asyncio.new_event_loop()
-        try:
-            loop.run_until_complete(client.aclose())
-        finally:
-            loop.close()
-    _backend_client = None
-
-
-def _initialise_backend(*, base_url: str | None = None) -> None:
-    """Configure service clients against an in-process or remote backend."""
-
-    global _backend_app, _backend_transport, _backend_client
-    global contract_service, dq_service, governance_service
-
-    _close_backend_client()
-
-    client_base_url = (base_url.rstrip("/") if base_url else "http://dc43-services")
-
-    if base_url:
-        _backend_app = None
-        _backend_transport = None
-        _backend_client = httpx.AsyncClient(base_url=client_base_url)
-    else:
-        _backend_app = build_local_app(store)
-        _backend_transport = ASGITransport(app=_backend_app)
-        _backend_client = httpx.AsyncClient(
-            transport=_backend_transport,
-            base_url=client_base_url,
-        )
-
-    contract_service = RemoteContractServiceClient(
-        base_url=client_base_url,
-        client=_backend_client,
-    )
-    dq_service = RemoteDataQualityServiceClient(
-        base_url=client_base_url,
-        client=_backend_client,
-    )
-    governance_service = RemoteGovernanceServiceClient(
-        base_url=client_base_url,
-        client=_backend_client,
-    )
-
-
-_initialise_backend(base_url=os.getenv("DC43_DEMO_BACKEND_URL"))
-
-
-def _wait_for_backend(base_url: str, timeout: float = 30.0) -> None:
-    """Block until the backend responds or ``timeout`` elapses."""
-
-    deadline = time.monotonic() + timeout
-    probe_url = f"{base_url.rstrip('/')}/openapi.json"
-    with httpx.Client(timeout=2.0) as client:
-        while True:
-            try:
-                response = client.get(probe_url)
-                if response.status_code < 500:
-                    return
-            except httpx.HTTPError:
-                pass
-            if time.monotonic() >= deadline:
-                raise RuntimeError(f"Backend at {base_url} failed to start within {timeout}s")
-            time.sleep(0.2)
-
-
-async def _expectation_predicates(contract: OpenDataContractStandard) -> Dict[str, str]:
-    plan = await asyncio.to_thread(dq_service.describe_expectations, contract=contract)
-    mapping: Dict[str, str] = {}
-    for item in plan:
-        key = item.get("key") if isinstance(item, Mapping) else None
-        predicate = item.get("predicate") if isinstance(item, Mapping) else None
-        if isinstance(key, str) and isinstance(predicate, str):
-            mapping[key] = predicate
-    return mapping
-
-# Populate server paths with sample datasets matching recorded versions
-_sample_records = json.loads((RECORDS_DIR / "datasets.json").read_text())
-for _r in _sample_records:
-    try:
-        _c = store.get(_r["contract_id"], _r["contract_version"])
-    except FileNotFoundError:
-        continue
-    _srv = (_c.servers or [None])[0]
-    if not _srv or not _srv.path:
-        continue
-    _dest = Path(_srv.path)
-    base = _dest.parent if _dest.suffix else _dest
-    if base.name == _r["dataset_name"]:
-        target_root = base / _r["dataset_version"]
-    else:
-        target_root = base / _r["dataset_name"] / _r["dataset_version"]
-    _base_dir = SAMPLE_DIR / "data" / _r["dataset_name"]
-    _src_dir = _base_dir / _r["dataset_version"]
-    if _src_dir.is_dir():
-        shutil.copytree(_src_dir, target_root, dirs_exist_ok=True)
-        continue
-    _src_file = SAMPLE_DIR / "data" / f"{_r['dataset_name']}.json"
-    if not _src_file.exists():
-        continue
-    target_root.mkdir(parents=True, exist_ok=True)
-    shutil.copy2(_src_file, target_root / _src_file.name)
-
-app = FastAPI(title="DC43 Demo")
-templates = Jinja2Templates(directory=str(BASE_DIR / "templates"))
-app.mount("/static", StaticFiles(directory=str(BASE_DIR / "static"), check_dir=False), name="static")
-
-
-@dataclass
-class DatasetRecord:
-    contract_id: str
-    contract_version: str
-    dataset_name: str = ""
-    dataset_version: str = ""
-    status: str = "unknown"
-    dq_details: Dict[str, Any] = field(default_factory=dict)
-    run_type: str = "infer"
-    violations: int = 0
-    reason: str = ""
-    draft_contract_version: str | None = None
-    scenario_key: str | None = None
-
-
-_STATUS_BADGES: Dict[str, str] = {
-    "kept": "bg-success",
-    "updated": "bg-primary",
-    "relaxed": "bg-warning text-dark",
-    "removed": "bg-danger",
-    "added": "bg-info text-dark",
-    "missing": "bg-secondary",
-    "error": "bg-danger",
-    "warning": "bg-warning text-dark",
-    "not_nullable": "bg-info text-dark",
-}
-
-
-_DQ_STATUS_BADGES: Dict[str, str] = {
-    "ok": "bg-success",
-    "warn": "bg-warning text-dark",
-    "block": "bg-danger",
-    "stale": "bg-secondary",
-    "unknown": "bg-secondary",
-}
-
-
-_CONTRACT_STATUS_BADGES: Dict[str, str] = {
-    "active": "bg-success",
-    "draft": "bg-warning text-dark",
-    "deprecated": "bg-secondary",
-}
-
-
-def _dq_version_records(
-    dataset_id: str,
-    *,
-    contract: Optional[OpenDataContractStandard] = None,
-    dataset_path: Optional[str] = None,
-    dataset_records: Optional[Iterable[DatasetRecord]] = None,
-) -> List[Dict[str, Any]]:
-    """Return version → status entries for the supplied dataset id.
-
-    ``dataset_records`` can be provided to scope compatibility information to
-    runs that were produced for a specific contract version. This ensures, for
-    example, that the compatibility matrix rendered for ``orders`` version
-    ``1.0.0`` does not surface the validation outcome that belongs to the
-    ``1.1.0`` contract.
-    """
-
-    records: List[Dict[str, Any]] = []
-    entries = _dq_status_entries(dataset_id)
-
-    scoped_versions: set[str] = set()
-    dataset_record_map: Dict[str, DatasetRecord] = {}
-    if dataset_records:
-        for record in dataset_records:
-            if not record.dataset_version:
-                continue
-            scoped_versions.add(record.dataset_version)
-            dataset_record_map[record.dataset_version] = record
-
-    dataset_dir = _dataset_root_for(dataset_id, dataset_path)
-    skip_fs_check = False
-    if contract and contract.servers:
-        server = contract.servers[0]
-        fmt = (getattr(server, "format", "") or "").lower()
-        if fmt == "delta":
-            skip_fs_check = True
-
-    seen_versions: set[str] = set()
-    for display_version, stored_version, payload in entries:
-        record = dataset_record_map.get(display_version)
-        payload_contract_id = str(payload.get("contract_id") or "")
-        payload_contract_version = str(payload.get("contract_version") or "")
-        if contract and (contract.id or contract.version):
-            contract_id_value = contract.id or ""
-            if payload_contract_id and payload_contract_version:
-                if (
-                    payload_contract_id != contract_id_value
-                    or payload_contract_version != contract.version
-                ):
-                    continue
-            elif scoped_versions and display_version not in scoped_versions:
-                continue
-        elif scoped_versions and display_version not in scoped_versions:
-            continue
-        if not skip_fs_check and dataset_dir is not None:
-            if not _has_version_materialisation(dataset_dir, display_version):
-                continue
-        status_value = str(payload.get("status", "unknown") or "unknown")
-        records.append(
-            {
-                "version": display_version,
-                "stored_version": stored_version,
-                "status": status_value,
-                "status_label": status_value.replace("_", " ").title(),
-                "badge": _DQ_STATUS_BADGES.get(status_value, "bg-secondary"),
-                "contract_id": payload_contract_id or (record.contract_id if record else ""),
-                "contract_version": payload_contract_version
-                or (record.contract_version if record else ""),
-                "recorded_at": payload.get("recorded_at"),
-            }
-        )
-        seen_versions.add(display_version)
-
-    # If we scoped by contract runs, surface any versions without a stored DQ
-    # payload using the dataset records so the UI can still display a verdict.
-    if scoped_versions:
-        for missing_version in scoped_versions - seen_versions:
-            record = dataset_record_map.get(missing_version)
-            status_value = str(record.status or "unknown") if record else "unknown"
-            records.append(
-                {
-                    "version": missing_version,
-                    "stored_version": _safe_fs_name(missing_version),
-                    "status": status_value,
-                    "status_label": status_value.replace("_", " ").title(),
-                    "badge": _DQ_STATUS_BADGES.get(status_value, "bg-secondary"),
-                    "contract_id": record.contract_id if record else "",
-                    "contract_version": record.contract_version if record else "",
-                    "recorded_at": None,
-                }
-            )
-
-    records.sort(key=lambda item: _version_sort_key(item["version"]))
-    return records
-
-
-def _server_details(contract: OpenDataContractStandard) -> Optional[Dict[str, Any]]:
-    """Summarise the first server entry for UI consumption."""
-
-    if not contract.servers:
-        return None
-    first = contract.servers[0]
-    custom: Dict[str, Any] = custom_properties_dict(first)
-    dataset_id = contract.id or getattr(first, "dataset", None) or contract.id
-    info: Dict[str, Any] = {
-        "server": getattr(first, "server", ""),
-        "type": getattr(first, "type", ""),
-        "format": getattr(first, "format", ""),
-        "path": getattr(first, "path", ""),
-        "dataset": getattr(first, "dataset", ""),
-        "dataset_id": dataset_id,
-    }
-    if custom:
-        info["custom"] = custom
-        if "dc43.versioning" in custom:
-            info["versioning"] = custom.get("dc43.versioning")
-        if "dc43.pathPattern" in custom:
-            info["path_pattern"] = custom.get("dc43.pathPattern")
-    return info
-
-
-def _format_scope(scope: str | None) -> str:
-    """Return a human readable label for change log scopes."""
-
-    if not scope or scope == "contract":
-        return "Contract"
-    if scope.startswith("field:"):
-        return f"Field {scope.split(':', 1)[1]}"
-    return scope.replace("_", " ").title()
-
-
-def _stringify_value(value: Any) -> str:
-    """Return a readable representation for rule parameter values."""
-
-    if isinstance(value, (list, tuple, set)):
-        return ", ".join(str(item) for item in value)
-    return str(value)
-
-
-def _quality_rule_summary(dq: DataQuality) -> Dict[str, Any]:
-    """Produce a structured summary for a data-quality rule."""
-
-    conditions: List[str] = []
-    if dq.description:
-        conditions.append(str(dq.description))
-
-    if dq.mustBeGreaterThan is not None:
-        conditions.append(f"Value must be greater than {dq.mustBeGreaterThan}")
-    if dq.mustBeGreaterOrEqualTo is not None:
-        conditions.append(f"Value must be greater than or equal to {dq.mustBeGreaterOrEqualTo}")
-    if dq.mustBeLessThan is not None:
-        conditions.append(f"Value must be less than {dq.mustBeLessThan}")
-    if dq.mustBeLessOrEqualTo is not None:
-        conditions.append(f"Value must be less than or equal to {dq.mustBeLessOrEqualTo}")
-    if dq.mustBeBetween:
-        low, high = dq.mustBeBetween
-        conditions.append(f"Value must be between {low} and {high}")
-    if dq.mustNotBeBetween:
-        low, high = dq.mustNotBeBetween
-        conditions.append(f"Value must not be between {low} and {high}")
-
-    if dq.mustBe is not None:
-        if (dq.rule or "").lower() == "regex":
-            conditions.append(f"Value must match the pattern {dq.mustBe}")
-        elif isinstance(dq.mustBe, (list, tuple, set)):
-            conditions.append(
-                "Value must be one of: " + ", ".join(str(item) for item in dq.mustBe)
-            )
-        else:
-            conditions.append(f"Value must be {_stringify_value(dq.mustBe)}")
-
-    if dq.mustNotBe is not None:
-        if isinstance(dq.mustNotBe, (list, tuple, set)):
-            conditions.append(
-                "Value must not be any of: "
-                + ", ".join(str(item) for item in dq.mustNotBe)
-            )
-        else:
-            conditions.append(f"Value must not be {_stringify_value(dq.mustNotBe)}")
-
-    if dq.query:
-        engine = (dq.engine or "spark_sql").replace("_", " ")
-        conditions.append(f"Query ({engine}): {dq.query}")
-
-    if not conditions:
-        label = dq.rule or dq.name or "rule"
-        conditions.append(f"See contract metadata for details on {label}.")
-
-    title = dq.name or dq.rule or "Rule"
-    title = title.replace("_", " ").title()
-
-    return {
-        "title": title,
-        "conditions": conditions,
-        "severity": dq.severity,
-        "dimension": dq.dimension,
-    }
-
-
-def _field_quality_sections(contract: OpenDataContractStandard) -> List[Dict[str, Any]]:
-    """Return quality rule summaries grouped per field."""
-
-    sections: List[Dict[str, Any]] = []
-    for obj in contract.schema_ or []:
-        for prop in obj.properties or []:
-            rules: List[Dict[str, Any]] = []
-            if prop.required:
-                rules.append(
-                    {
-                        "title": "Required",
-                        "conditions": [
-                            "Field must always be present (non-null values required)."
-                        ],
-                    }
-                )
-            if prop.unique:
-                rules.append(
-                    {
-                        "title": "Unique",
-                        "conditions": [
-                            "Each record must contain a distinct value for this field.",
-                        ],
-                    }
-                )
-            for dq in prop.quality or []:
-                rules.append(_quality_rule_summary(dq))
-
-            sections.append(
-                {
-                    "name": prop.name or "",
-                    "type": prop.physicalType or "",
-                    "required": bool(prop.required),
-                    "rules": rules,
-                }
-            )
-    return sections
-
-
-def _dataset_quality_sections(contract: OpenDataContractStandard) -> List[Dict[str, Any]]:
-    """Return dataset-level quality rules defined on schema objects."""
-
-    sections: List[Dict[str, Any]] = []
-    for obj in contract.schema_ or []:
-        rules = [_quality_rule_summary(dq) for dq in obj.quality or []]
-        if rules:
-            sections.append({"name": obj.name or contract.id or "dataset", "rules": rules})
-    return sections
-
-
-def _summarise_change_entry(entry: Mapping[str, Any]) -> str:
-    details = entry.get("details")
-    if isinstance(details, Mapping):
-        for key in ("message", "reason"):
-            message = details.get(key)
-            if message:
-                return str(message)
-    target = entry.get("constraint") or entry.get("rule") or entry.get("kind")
-    status = entry.get("status")
-    if target and status:
-        return f"{str(target).replace('_', ' ').title()} {str(status).replace('_', ' ')}."
-    if status:
-        return str(status).replace("_", " ").title()
-    return ""
-
-
-def _contract_change_log(contract: OpenDataContractStandard) -> List[Dict[str, Any]]:
-    """Extract change log entries from the contract custom properties."""
-
-    entries: List[Dict[str, Any]] = []
-    for prop in normalise_custom_properties(contract.customProperties):
-        if isinstance(prop, Mapping):
-            key = prop.get("property")
-            value = prop.get("value")
-        else:
-            key = getattr(prop, "property", None)
-            value = getattr(prop, "value", None)
-        if key != "draft_change_log":
-            continue
-        try:
-            items = list(value or [])
-        except TypeError:
-            continue
-        for item in items:
-            if not isinstance(item, Mapping):
-                continue
-            details = item.get("details")
-            details_text = ""
-            if details is not None:
-                try:
-                    details_text = json.dumps(details, indent=2, sort_keys=True, default=str)
-                except TypeError:
-                    details_text = str(details)
-            status = str(item.get("status", ""))
-            entries.append(
-                {
-                    "scope": item.get("scope", ""),
-                    "scope_label": _format_scope(item.get("scope")),
-                    "kind": item.get("kind", ""),
-                    "status": status,
-                    "status_label": status.replace("_", " ").title(),
-                    "constraint": item.get("constraint"),
-                    "rule": item.get("rule"),
-                    "summary": _summarise_change_entry(item),
-                    "details_text": details_text,
-                }
-            )
-        break
-    return entries
-
-
-def load_records() -> List[DatasetRecord]:
-    raw = json.loads(DATASETS_FILE.read_text())
-    return [DatasetRecord(**r) for r in raw]
-
-
-def save_records(records: List[DatasetRecord]) -> None:
-    DATASETS_FILE.write_text(
-        json.dumps([r.__dict__ for r in records], indent=2), encoding="utf-8"
-    )
-
-
-# Default slice activations used to drive the ``latest`` aliases for scenarios.
-_DEFAULT_SLICE = {
-    "orders": "2024-01-01",
-    "customers": "2024-01-01",
-}
-
-_INVALID_SLICE = {
-    "orders": "2025-09-28",
-    "orders__valid": "2025-09-28",
-    "orders__reject": "2025-09-28",
-    "customers": "2024-01-01",
-}
-
-# Predefined pipeline scenarios exposed in the UI. Each scenario describes the
-# parameters passed to the example pipeline along with a human readable
-# description shown to the user.
-SCENARIOS: Dict[str, Dict[str, Any]] = {
-    "no-contract": {
-        "label": "No contract provided",
-        "description": (
-            "<p>Run the pipeline without supplying an output contract.</p>"
-            "<ul>"
-            "<li><strong>Inputs:</strong> Reads <code>orders:1.1.0</code> and "
-            "<code>customers:1.0.0</code> with schema validation.</li>"
-            "<li><strong>Contract:</strong> None provided, so no draft can be"
-            " created.</li>"
-            "<li><strong>Writes:</strong> Planned dataset <code>result-no-existing-contract</code>"
-            " is blocked before any files are materialised, so no version is"
-            " assigned.</li>"
-            "<li><strong>Status:</strong> The run exits with an error because the contract is"
-            " missing.</li>"
-            "</ul>"
-        ),
-        "diagram": (
-            "<div class=\"mermaid\">"
-            + dedent(
-                """
-                flowchart TD
-                    Orders["orders latest → 2024-01-01\ncontract orders:1.1.0"] --> Join[Join datasets]
-                    Customers["customers latest → 2024-01-01\ncontract customers:1.0.0"] --> Join
-                    Join --> Write["Plan result-no-existing-contract\nno output contract"]
-                    Write -->|no contract| Block[Run blocked, nothing written]
-                """
-            ).strip()
-            + "</div>"
-        ),
-        "activate_versions": dict(_DEFAULT_SLICE),
-        "params": {
-            "contract_id": None,
-            "contract_version": None,
-            "dataset_name": "result-no-existing-contract",
-            "run_type": "enforce",
-        },
-    },
-    "ok": {
-        "label": "Existing contract OK",
-        "description": (
-            "<p>Happy path using contract <code>orders_enriched:1.0.0</code>.</p>"
-            "<ul>"
-            "<li><strong>Inputs:</strong> Reads <code>orders:1.1.0</code> and"
-            " <code>customers:1.0.0</code> then aligns to the target schema.</li>"
-            "<li><strong>Contract:</strong> Targets <code>orders_enriched:1.0.0</code>"
-            " with no draft changes.</li>"
-            "<li><strong>Writes:</strong> Persists dataset <code>orders_enriched</code>"
-            " tagged with the run timestamp so repeated runs never collide.</li>"
-            "<li><strong>Status:</strong> Post-write validation reports OK.</li>"
-            "</ul>"
-        ),
-        "diagram": (
-            "<div class=\"mermaid\">"
-            + dedent(
-                """
-                flowchart TD
-                    Orders["orders latest → 2024-01-01\ncontract orders:1.1.0"] --> Join[Join datasets]
-                    Customers["customers latest → 2024-01-01\ncontract customers:1.0.0"] --> Join
-                    Join --> Validate[Align to contract orders_enriched:1.0.0]
-                    Validate --> Write["orders_enriched «timestamp»\ncontract orders_enriched:1.0.0"]
-                    Write --> Status[Run status: OK]
-                """
-            ).strip()
-            + "</div>"
-        ),
-        "activate_versions": dict(_DEFAULT_SLICE),
-        "params": {
-            "contract_id": "orders_enriched",
-            "contract_version": "1.0.0",
-            "run_type": "enforce",
-        },
-    },
-    "dq": {
-        "label": "Existing contract fails DQ",
-        "description": (
-            "<p>Demonstrates a data quality failure.</p>"
-            "<ul>"
-            "<li><strong>Inputs:</strong> Reads <code>orders:1.1.0</code> and"
-            " <code>customers:1.0.0</code>.</li>"
-            "<li><strong>Contract:</strong> Validates against"
-            " <code>orders_enriched:1.1.0</code> and prepares draft"
-            " <code>orders_enriched:1.2.0</code>.</li>"
-            "<li><strong>Writes:</strong> Persists"
-            " <code>orders_enriched</code> with the run timestamp before"
-            " governance flips the outcome to <code>block</code> and records"
-            " draft <code>orders_enriched:1.2.0</code>.</li>"
-            "<li><strong>Status:</strong> The enforcement run errors when rule"
-            " <code>amount &gt; 100</code> is violated.</li>"
-            "</ul>"
-        ),
-        "diagram": (
-            "<div class=\"mermaid\">"
-            + dedent(
-                """
-                flowchart TD
-                    Orders["orders latest → 2024-01-01\ncontract orders:1.1.0"] --> Join[Join datasets]
-                    Customers["customers latest → 2024-01-01\ncontract customers:1.0.0"] --> Join
-                    Join --> Write["orders_enriched «timestamp»\ncontract orders_enriched:1.1.0"]
-                    Write --> Governance[Post-write validation]
-                    Governance --> Draft[Draft orders_enriched 1.2.0]
-                    Governance -->|violations| Block["DQ verdict: block"]
-                """
-            ).strip()
-            + "</div>"
-        ),
-        "activate_versions": dict(_DEFAULT_SLICE),
-        "params": {
-            "contract_id": "orders_enriched",
-            "contract_version": "1.1.0",
-            "run_type": "enforce",
-            "collect_examples": True,
-            "examples_limit": 3,
-        },
-    },
-    "schema-dq": {
-        "label": "Contract fails schema and DQ",
-        "description": (
-            "<p>Shows combined schema and data quality issues.</p>"
-            "<ul>"
-            "<li><strong>Inputs:</strong> Reads <code>orders:1.1.0</code> and"
-            " <code>customers:1.0.0</code>.</li>"
-            "<li><strong>Contract:</strong> Targets <code>orders_enriched:2.0.0</code>"
-            " and proposes draft <code>orders_enriched:2.1.0</code>.</li>"
-            "<li><strong>Writes:</strong> Persists"
-            " <code>orders_enriched</code> with the run timestamp, then"
-            " validation downgrades the outcome to <code>block</code> while"
-            " recording draft <code>orders_enriched:2.1.0</code>.</li>"
-            "<li><strong>Status:</strong> Schema drift plus failed expectations"
-            " produce an error outcome.</li>"
-            "</ul>"
-        ),
-        "diagram": (
-            "<div class=\"mermaid\">"
-            + dedent(
-                """
-                flowchart TD
-                    Orders["orders latest → 2024-01-01\ncontract orders:1.1.0"] --> Join[Join datasets]
-                    Customers["customers latest → 2024-01-01\ncontract customers:1.0.0"] --> Join
-                    Join --> Align[Schema align to contract orders_enriched:2.0.0]
-                    Align --> Write["orders_enriched «timestamp»\ncontract orders_enriched:2.0.0"]
-                    Write --> Governance[Post-write validation]
-                    Governance --> Draft[Draft orders_enriched 2.1.0]
-                    Governance -->|violations| Block["DQ verdict: block"]
-                """
-            ).strip()
-            + "</div>"
-        ),
-        "activate_versions": dict(_DEFAULT_SLICE),
-        "params": {
-            "contract_id": "orders_enriched",
-            "contract_version": "2.0.0",
-            "run_type": "enforce",
-        },
-    },
-    "read-invalid-block": {
-        "label": "Invalid input blocked",
-        "description": (
-            "<p>Attempts to process the latest slice (→2025-09-28) flagged as invalid.</p>"
-            "<ul>"
-            "<li><strong>Inputs:</strong> Governance records mark"
-            " <code>orders latest → 2025-09-28</code> as <code>block</code> while pointing"
-            " at curated <code>valid</code> and <code>reject</code> slices.</li>"
-            "<li><strong>Contract:</strong> Targets <code>orders_enriched:1.1.0</code>"
-            " but enforcement aborts before writes.</li>"
-            "<li><strong>Outputs:</strong> None; the job fails fast.</li>"
-            "<li><strong>Governance:</strong> Stub DQ client returns the stored"
-            " `block` verdict and its auxiliary dataset hints.</li>"
-            "</ul>"
-        ),
-        "diagram": (
-            "<div class=\"mermaid\">"
-            + dedent(
-                """
-                flowchart TD
-                    Invalid["orders latest → 2025-09-28\ncontract orders:1.1.0\nDQ status: block"] -->|default enforcement| Halt[Read aborted]
-                    Invalid -.-> Valid["orders::valid latest__valid → 2025-09-28\ncontract orders:1.1.0"]
-                    Invalid -.-> Reject["orders::reject latest__reject → 2025-09-28\ncontract orders:1.1.0"]
-                """
-            ).strip()
-            + "</div>"
-        ),
-        "activate_versions": dict(_INVALID_SLICE),
-        "params": {
-            "contract_id": "orders_enriched",
-            "contract_version": "1.1.0",
-            "run_type": "enforce",
-            "inputs": {
-                "orders": {
-                    "dataset_version": "latest",
-                }
-            },
-        },
-    },
-    "read-valid-subset": {
-        "label": "Prefer valid subset",
-        "description": (
-            "<p>Steers reads toward the curated valid slice.</p>"
-            "<ul>"
-            "<li><strong>Inputs:</strong> Uses <code>orders::valid</code>"
-            " <code>latest__valid → 2025-09-28</code> alongside"
-            " <code>customers latest → 2024-01-01</code> to satisfy governance.</li>"
-            "<li><strong>Contract:</strong> Applies <code>orders_enriched:1.1.0</code>"
-            " and keeps draft creation disabled.</li>"
-            "<li><strong>Outputs:</strong> Writes <code>orders_enriched</code>"
-            " stamped with the run timestamp under contract"
-            " <code>orders_enriched:1.1.0</code> with a clean DQ verdict.</li>"
-            "<li><strong>Governance:</strong> Stub evaluates post-write metrics"
-            " and records an OK status.</li>"
-            "</ul>"
-        ),
-        "diagram": (
-            "<div class=\"mermaid\">"
-            + dedent(
-                """
-                flowchart TD
-                    Valid["orders::valid latest__valid → 2025-09-28\ncontract orders:1.1.0"] --> Join[Join datasets]
-                    Customers["customers latest → 2024-01-01\ncontract customers:1.0.0"] --> Join
-                    Join --> Write["orders_enriched «timestamp»\ncontract orders_enriched:1.1.0"]
-                    Write --> Governance[Governance verdict ok]
-                    Governance --> Status["DQ status: ok"]
-                """
-            ).strip()
-            + "</div>"
-        ),
-        "activate_versions": dict(_INVALID_SLICE),
-        "params": {
-            "contract_id": "orders_enriched",
-            "contract_version": "1.1.0",
-            "run_type": "observe",
-            "collect_examples": True,
-            "examples_limit": 3,
-            "inputs": {
-                "orders": {
-                    "dataset_id": "orders::valid",
-                    "dataset_version": "latest__valid",
-                }
-            },
-        },
-    },
-    "read-valid-subset-violation": {
-        "label": "Valid subset, invalid output",
-        "description": (
-            "<p>Highlights when clean inputs still breach the output contract.</p>"
-            "<ul>"
-            "<li><strong>Inputs:</strong> Same curated"
-            " <code>orders::valid</code> <code>latest__valid → 2025-09-28</code> slice.</li>"
-            "<li><strong>Contract:</strong> Writes to"
-            " <code>orders_enriched</code> under <code>orders_enriched:1.1.0</code>.</li>"
-            "<li><strong>Outputs:</strong> Produces <code>orders_enriched</code>"
-            " (timestamped under contract <code>1.1.0</code>) but post-write checks fail because"
-            " the demo purposely lowers one amount below the"
-            " <code>&gt; 100</code> expectation.</li>"
-            "<li><strong>Governance:</strong> Stub DQ client records a blocking"
-            " verdict and drafts <code>orders_enriched:1.2.0</code>.</li>"
-            "</ul>"
-        ),
-        "diagram": (
-            "<div class=\"mermaid\">"
-            + dedent(
-                """
-                flowchart TD
-                    Valid["orders::valid latest__valid → 2025-09-28\ncontract orders:1.1.0"] --> Join[Join datasets]
-                    Join --> Adjust[Lower amount to 60]
-                    Adjust --> Write["orders_enriched «timestamp»\ncontract orders_enriched:1.1.0"]
-                    Write --> Governance[Governance verdict block]
-                    Governance --> Draft["Draft orders_enriched 1.2.0"]
-                    Governance --> Status["DQ status: block"]
-                """
-            ).strip()
-            + "</div>"
-        ),
-        "activate_versions": dict(_INVALID_SLICE),
-        "params": {
-            "contract_id": "orders_enriched",
-            "contract_version": "1.1.0",
-            "run_type": "enforce",
-            "collect_examples": True,
-            "examples_limit": 3,
-            "inputs": {
-                "orders": {
-                    "dataset_id": "orders::valid",
-                    "dataset_version": "latest__valid",
-                }
-            },
-            "output_adjustment": "valid-subset-violation",
-        },
-    },
-    "read-override-full": {
-        "label": "Force blocked slice (manual override)",
-        "description": (
-            "<p>Documents what happens when the blocked data is forced through.</p>"
-            "<ul>"
-            "<li><strong>Inputs:</strong> Reuses the blocked"
-            " <code>orders latest → 2025-09-28</code> and downgrades the read status to"
-            " <code>warn</code>.</li>"
-            "<li><strong>Override strategy:</strong> Uses"
-            " <code>allow-block</code> to document that the blocked slice was"
-            " manually forced through despite the governance verdict.</li>"
-            "<li><strong>Contract:</strong> Applies"
-            " <code>orders_enriched:1.1.0</code> and captures draft"
-            " <code>orders_enriched:1.2.0</code>.</li>"
-            "<li><strong>Outputs:</strong> Writes <code>orders_enriched</code>"
-            " (timestamped under contract <code>1.1.0</code>) while surfacing the manual override"
-            " note alongside the reject-row metrics.</li>"
-            "<li><strong>Governance:</strong> Stub records the downgrade in the"
-            " run summary alongside violation counts and the explicit override"
-            " note.</li>"
-            "</ul>"
-        ),
-        "diagram": (
-            "<div class=\"mermaid\">"
-            + dedent(
-                """
-                flowchart TD
-                    Invalid["orders latest → 2025-09-28\ncontract orders:1.1.0\nDQ status: block"] --> Override[Downgrade to warn]
-                    Override --> Write["orders_enriched «timestamp»\ncontract orders_enriched:1.1.0"]
-                    Write --> Governance[Governance verdict warn]
-                    Governance --> Draft["Draft orders_enriched 1.2.0"]
-                    Governance --> Status["DQ status: warn"]
-                """
-            ).strip()
-            + "</div>"
-        ),
-        "activate_versions": dict(_INVALID_SLICE),
-        "params": {
-            "contract_id": "orders_enriched",
-            "contract_version": "1.1.0",
-            "run_type": "observe",
-            "collect_examples": True,
-            "examples_limit": 3,
-            "inputs": {
-                "orders": {
-                    "dataset_version": "latest",
-                    "status_strategy": {
-                        "name": "allow-block",
-                        "note": "Manual override: forced latest slice (→2025-09-28)",
-                        "target_status": "warn",
-                    },
-                }
-            },
-            "output_adjustment": "amplify-negative",
-        },
-    },
-    "split-lenient": {
-        "label": "Split invalid rows",
-        "description": (
-            "<p>Routes violations to dedicated datasets using the split strategy.</p>"
-            "<ul>"
-            "<li><strong>Inputs:</strong> Reads <code>orders:1.1.0</code> and"
-            " <code>customers:1.0.0</code> before aligning to"
-            " <code>orders_enriched:1.1.0</code>.</li>"
-            "<li><strong>Contract:</strong> Validates against"
-            " <code>orders_enriched:1.1.0</code> and stores draft"
-            " <code>orders_enriched:1.2.0</code> when rejects exist.</li>"
-            "<li><strong>Writes:</strong> Persists three datasets sharing the same"
-            " timestamp: the contracted"
-            " <code>orders_enriched</code> (full slice),"
-            " <code>orders_enriched::valid</code>, and"
-            " <code>orders_enriched::reject</code>.</li>"
-            "<li><strong>Status:</strong> Run finishes with a warning because"
-            " validation finds violations, and the UI links the auxiliary"
-            " datasets.</li>"
-            "</ul>"
-        ),
-        "diagram": (
-            "<div class=\"mermaid\">"
-            + dedent(
-                """
-                flowchart TD
-                    Orders["orders latest → 2024-01-01\ncontract orders:1.1.0"] --> Join[Join datasets]
-                    Customers["customers latest → 2024-01-01\ncontract customers:1.0.0"] --> Join
-                    Join --> Validate[Validate contract orders_enriched:1.1.0]
-                    Validate --> Strategy[Split strategy]
-                    Strategy --> Full["orders_enriched «timestamp»\ncontract orders_enriched:1.1.0"]
-                    Strategy --> Valid["orders_enriched::valid «timestamp»\ncontract orders_enriched:1.1.0"]
-                    Strategy --> Reject["orders_enriched::reject «timestamp»\ncontract orders_enriched:1.1.0"]
-                """
-            ).strip()
-            + "</div>"
-        ),
-        "activate_versions": dict(_DEFAULT_SLICE),
-        "params": {
-            "contract_id": "orders_enriched",
-            "contract_version": "1.1.0",
-            "run_type": "observe",
-            "collect_examples": True,
-            "examples_limit": 3,
-            "violation_strategy": {
-                "name": "split",
-                "include_valid": True,
-                "include_reject": True,
-                "write_primary_on_violation": True,
-            },
-        },
-    },
-}
-
-
-def _scenario_dataset_name(params: Mapping[str, Any]) -> str:
-    """Return the expected output dataset for a scenario."""
-
-    dataset_name = params.get("dataset_name")
-    if dataset_name:
-        return str(dataset_name)
-    contract_id = params.get("contract_id")
-    if contract_id:
-        return str(contract_id)
-    dataset_id = params.get("dataset_id")
-    if dataset_id:
-        return str(dataset_id)
-    return "result"
-
-
-def scenario_run_rows(records: Iterable[DatasetRecord]) -> List[Dict[str, Any]]:
-    """Return scenario metadata enriched with the latest recorded run."""
-
-    by_dataset: Dict[str, List[DatasetRecord]] = {}
-    by_scenario: Dict[str, List[DatasetRecord]] = {}
-    for record in records:
-        if record.dataset_name:
-            by_dataset.setdefault(record.dataset_name, []).append(record)
-        if record.scenario_key:
-            by_scenario.setdefault(record.scenario_key, []).append(record)
-
-    for entries in by_dataset.values():
-        entries.sort(key=lambda item: _version_sort_key(item.dataset_version or ""))
-    for entries in by_scenario.values():
-        entries.sort(key=lambda item: _version_sort_key(item.dataset_version or ""))
-
-    rows: List[Dict[str, Any]] = []
-    for key, cfg in SCENARIOS.items():
-        params: Mapping[str, Any] = cfg.get("params", {})
-        dataset_name = _scenario_dataset_name(params)
-        dataset_records: List[DatasetRecord] = list(by_scenario.get(key, []))
-
-        if not dataset_records:
-            candidate_records = by_dataset.get(dataset_name, [])
-            if candidate_records:
-                contract_id = params.get("contract_id")
-                contract_version = params.get("contract_version")
-                run_type = params.get("run_type")
-                filtered: List[DatasetRecord] = []
-                for record in candidate_records:
-                    if record.scenario_key:
-                        continue
-                    if contract_id and record.contract_id and record.contract_id != contract_id:
-                        continue
-                    if (
-                        contract_version
-                        and record.contract_version
-                        and record.contract_version != contract_version
-                    ):
-                        continue
-                    if run_type and record.run_type and record.run_type != run_type:
-                        continue
-                    filtered.append(record)
-                if filtered:
-                    dataset_records = filtered
-                else:
-                    dataset_records = [rec for rec in candidate_records if not rec.scenario_key]
-
-        dataset_records = list(dataset_records)
-        dataset_records.sort(key=lambda item: _version_sort_key(item.dataset_version or ""))
-        latest_record = dataset_records[-1] if dataset_records else None
-
-        rows.append(
-            {
-                "key": key,
-                "label": cfg.get("label", key.replace("-", " ").title()),
-                "description": cfg.get("description"),
-                "diagram": cfg.get("diagram"),
-                "dataset_name": dataset_name,
-                "contract_id": params.get("contract_id"),
-                "contract_version": params.get("contract_version"),
-                "run_type": params.get("run_type", "infer"),
-                "run_count": len(dataset_records),
-                "latest": latest_record.__dict__.copy() if latest_record else None,
-            }
-        )
-
-    return rows
-
-
-_FLASH_LOCK = Lock()
-_FLASH_MESSAGES: Dict[str, Dict[str, str | None]] = {}
-
-
-def queue_flash(message: str | None = None, error: str | None = None) -> str:
-    """Store a transient flash payload and return a lookup token."""
-
-    token = uuid4().hex
-    with _FLASH_LOCK:
-        _FLASH_MESSAGES[token] = {"message": message, "error": error}
-    return token
-
-
-def pop_flash(token: str) -> Tuple[str | None, str | None]:
-    """Return and remove the flash payload associated with ``token``."""
-
-    with _FLASH_LOCK:
-        payload = _FLASH_MESSAGES.pop(token, None) or {}
-    return payload.get("message"), payload.get("error")
-
-
-def load_contract_meta() -> List[Dict[str, Any]]:
-    """Return contract info derived from the store without extra metadata."""
-    meta: List[Dict[str, Any]] = []
-    for cid in store.list_contracts():
-        for ver in store.list_versions(cid):
-            try:
-                contract = store.get(cid, ver)
-            except FileNotFoundError:
-                continue
-            server = (contract.servers or [None])[0]
-            path = ""
-            if server:
-                parts: List[str] = []
-                if getattr(server, "path", None):
-                    parts.append(server.path)
-                if getattr(server, "dataset", None):
-                    parts.append(server.dataset)
-                path = "/".join(parts)
-            meta.append({"id": cid, "version": ver, "path": path})
-    return meta
-
-
-def save_contract_meta(meta: List[Dict[str, Any]]) -> None:
-    """No-op retained for backwards compatibility."""
-    return None
-
-
-def contract_to_dict(c: OpenDataContractStandard) -> Dict[str, Any]:
-    """Return a plain dict for a contract using public field aliases."""
-    try:
-        return c.model_dump(by_alias=True, exclude_none=True)
-    except AttributeError:  # pragma: no cover - Pydantic v1 fallback
-        return c.dict(by_alias=True, exclude_none=True)  # type: ignore[call-arg]
-
-
-@app.get("/api/contracts")
-async def api_contracts() -> List[Dict[str, Any]]:
-    return load_contract_meta()
-
-
-@app.get("/api/contracts/{cid}/{ver}")
-async def api_contract_detail(cid: str, ver: str) -> Dict[str, Any]:
-    try:
-        contract = store.get(cid, ver)
-    except FileNotFoundError as exc:
-        raise HTTPException(status_code=404, detail=str(exc))
-    datasets = [r.__dict__ for r in load_records() if r.contract_id == cid and r.contract_version == ver]
-    expectations = await _expectation_predicates(contract)
-    return {
-        "contract": contract_to_dict(contract),
-        "datasets": datasets,
-        "expectations": expectations,
-    }
-
-
-@app.get("/api/contracts/{cid}/{ver}/preview")
-async def api_contract_preview(
-    cid: str,
-    ver: str,
-    dataset_version: Optional[str] = None,
-    dataset_id: Optional[str] = None,
-    limit: int = 100,
-) -> Dict[str, Any]:
-    if read_with_contract is None or ContractVersionLocator is None:
-        raise HTTPException(status_code=503, detail="pyspark is required for data previews")
-    try:
-        contract = store.get(cid, ver)
-    except FileNotFoundError as exc:
-        raise HTTPException(status_code=404, detail=str(exc))
-
-    effective_dataset_id = str(dataset_id or contract.id or cid)
-    server = (contract.servers or [None])[0]
-    dataset_path_hint = getattr(server, "path", None) if server else None
-    version_contract = contract if effective_dataset_id == (contract.id or cid) else None
-    scoped_records = [
-        record
-        for record in load_records()
-        if record.contract_id == cid
-        and record.contract_version == ver
-        and record.dataset_name == effective_dataset_id
+CONTRACTS_TEMPLATE_DIR = contracts_server.BASE_DIR / "templates"
+PIPELINE_TEMPLATE_DIR = BASE_DIR / "templates"
+
+_env_loader = ChoiceLoader(
+    [
+        FileSystemLoader(str(PIPELINE_TEMPLATE_DIR)),
+        FileSystemLoader(str(CONTRACTS_TEMPLATE_DIR)),
     ]
-    version_records = _dq_version_records(
-        effective_dataset_id,
-        contract=version_contract,
-        dataset_path=dataset_path_hint if version_contract else None,
-        dataset_records=scoped_records,
-    )
-    known_versions = [entry["version"] for entry in version_records]
-    if not known_versions:
-        known_versions = ["latest"]
-    selected_version = str(dataset_version or known_versions[-1])
-    if selected_version not in known_versions:
-        known_versions = _sort_versions([*known_versions, selected_version])
-    limit = max(1, min(limit, 500))
-
-    try:
-        def _load_preview() -> tuple[list[Mapping[str, Any]], list[str]]:
-            spark = _spark_session()
-            locator = ContractVersionLocator(
-                dataset_version=selected_version,
-                dataset_id=effective_dataset_id,
-            )
-            df = read_with_contract(  # type: ignore[misc]
-                spark,
-                contract_id=cid,
-                contract_service=contract_service,
-                expected_contract_version=f"=={ver}",
-                dataset_locator=locator,
-                enforce=False,
-                auto_cast=False,
-                data_quality_service=dq_service,
-                return_status=False,
-            )
-            rows_raw = [row.asDict(recursive=True) for row in df.limit(limit).collect()]
-            return rows_raw, list(df.columns)
-
-        rows_raw, columns = await asyncio.to_thread(_load_preview)
-        rows = jsonable_encoder(rows_raw)
-    except Exception as exc:  # pragma: no cover - defensive guard for preview errors
-        raise HTTPException(status_code=500, detail=str(exc))
-
-    status_payload = _dq_status_payload(effective_dataset_id, selected_version)
-    status_value = str(status_payload.get("status", "unknown")) if status_payload else "unknown"
-    response = {
-        "dataset_id": effective_dataset_id,
-        "dataset_version": selected_version,
-        "rows": rows,
-        "columns": columns,
-        "limit": limit,
-        "known_versions": known_versions,
-        "status": {
-            "status": status_value,
-            "status_label": status_value.replace("_", " ").title(),
-            "badge": _DQ_STATUS_BADGES.get(status_value, "bg-secondary"),
-            "details": status_payload.get("details") if status_payload else None,
-        },
-    }
-    return response
-
-
-@app.post("/api/contracts/{cid}/{ver}/validate")
-async def api_validate_contract(cid: str, ver: str) -> Dict[str, str]:
-    return {"status": "active"}
-
-
-@app.get("/api/datasets")
-async def api_datasets() -> List[Dict[str, Any]]:
-    records = load_records()
-    return [r.__dict__.copy() for r in records]
-
-
-@app.get("/api/datasets/{dataset_version}")
-async def api_dataset_detail(dataset_version: str) -> Dict[str, Any]:
-    for r in load_records():
-        if r.dataset_version == dataset_version:
-            contract = store.get(r.contract_id, r.contract_version)
-            return {
-                "record": r.__dict__,
-                "contract": contract_to_dict(contract),
-                "expectations": await _expectation_predicates(contract),
-            }
-    raise HTTPException(status_code=404, detail="Dataset not found")
-
-
-@app.get("/", response_class=HTMLResponse)
-async def index(request: Request) -> HTMLResponse:
-    return templates.TemplateResponse("index.html", {"request": request})
-
-
-@app.get("/contracts", response_class=HTMLResponse)
-async def list_contracts(request: Request) -> HTMLResponse:
-    contract_ids = store.list_contracts()
-    return templates.TemplateResponse(
-        "contracts.html", {"request": request, "contracts": contract_ids}
-    )
-
-
-@app.get("/contracts/new", response_class=HTMLResponse)
-async def new_contract_form(request: Request) -> HTMLResponse:
-    editor_state = _contract_editor_state()
-    editor_state["version"] = editor_state.get("version") or "1.0.0"
-    context = _editor_context(request, editor_state=editor_state)
-    return templates.TemplateResponse("new_contract.html", context)
-
-
-@app.post("/contracts/new", response_class=HTMLResponse)
-async def create_contract(
-    request: Request,
-    payload: str = Form(...),
-) -> HTMLResponse:
-    error: Optional[str] = None
-    try:
-        editor_state = json.loads(payload)
-    except json.JSONDecodeError as exc:
-        error = f"Invalid editor payload: {exc.msg}"
-        editor_state = _contract_editor_state()
-    else:
-        try:
-            _validate_contract_payload(editor_state, editing=False)
-            model = _build_contract_from_payload(editor_state)
-            store.put(model)
-            return RedirectResponse(url="/contracts", status_code=303)
-        except (ValidationError, ValueError) as exc:
-            error = str(exc)
-        except Exception as exc:  # pragma: no cover - display unexpected errors
-            error = str(exc)
-    context = _editor_context(
-        request,
-        editor_state=editor_state,
-        error=error,
-    )
-    return templates.TemplateResponse("new_contract.html", context)
-
-
-
-
-@app.get("/contracts/{cid}", response_class=HTMLResponse)
-async def list_contract_versions(request: Request, cid: str) -> HTMLResponse:
-    versions = store.list_versions(cid)
-    if not versions:
-        raise HTTPException(status_code=404, detail="Contract not found")
-    records_by_version: Dict[str, List[DatasetRecord]] = {}
-    for record in load_records():
-        if record.contract_id != cid:
-            continue
-        records_by_version.setdefault(record.contract_version, []).append(record)
-
-    contracts = []
-    for ver in versions:
-        try:
-            contract = store.get(cid, ver)
-        except FileNotFoundError:
-            continue
-
-        status_raw = getattr(contract, "status", "") or "unknown"
-        status_value = str(status_raw).lower()
-        status_label = str(status_raw).replace("_", " ").title()
-        status_badge = _CONTRACT_STATUS_BADGES.get(status_value, "bg-secondary")
-
-        server_info = _server_details(contract)
-        dataset_hint = (
-            server_info.get("dataset_id")
-            if server_info
-            else (contract.id or cid)
-        )
-
-        latest_run: Optional[DatasetRecord] = None
-        run_entries = records_by_version.get(ver, [])
-        if run_entries:
-            run_entries.sort(key=lambda item: _version_sort_key(item.dataset_version or ""))
-            latest_run = run_entries[-1]
-
-        contracts.append(
-            {
-                "id": cid,
-                "version": ver,
-                "status": status_value,
-                "status_label": status_label,
-                "status_badge": status_badge,
-                "server": server_info,
-                "dataset_hint": dataset_hint,
-                "latest_run": latest_run.__dict__ if latest_run else None,
-            }
-        )
-    context = {"request": request, "contract_id": cid, "contracts": contracts}
-    return templates.TemplateResponse("contract_versions.html", context)
-
-
-@app.get("/contracts/{cid}/{ver}", response_class=HTMLResponse)
-async def contract_detail(request: Request, cid: str, ver: str) -> HTMLResponse:
-    try:
-        contract = store.get(cid, ver)
-    except FileNotFoundError as exc:
-        raise HTTPException(status_code=404, detail=str(exc))
-    datasets = [r for r in load_records() if r.contract_id == cid and r.contract_version == ver]
-    field_quality = _field_quality_sections(contract)
-    dataset_quality = _dataset_quality_sections(contract)
-    change_log = _contract_change_log(contract)
-    server_info = _server_details(contract)
-    dataset_id = server_info.get("dataset_id") if server_info else contract.id or cid
-    dataset_path_hint = server_info.get("path") if server_info else None
-    version_records = _dq_version_records(
-        dataset_id or cid,
-        contract=contract,
-        dataset_path=dataset_path_hint,
-        dataset_records=datasets,
-    )
-    version_list = [entry["version"] for entry in version_records]
-    status_map = {
-        entry["version"]: {
-            "status": entry["status"],
-            "label": entry["status_label"],
-            "badge": entry["badge"],
-        }
-        for entry in version_records
-    }
-    default_index = len(version_list) - 1 if version_list else None
-    context = {
-        "request": request,
-        "contract": contract_to_dict(contract),
-        "datasets": datasets,
-        "expectations": await _expectation_predicates(contract),
-        "field_quality": field_quality,
-        "dataset_quality": dataset_quality,
-        "change_log": change_log,
-        "status_badges": _STATUS_BADGES,
-        "server_info": server_info,
-        "compatibility_versions": version_records,
-        "preview_versions": version_list,
-        "preview_status_map": status_map,
-        "preview_default_index": default_index,
-        "preview_dataset_id": dataset_id,
-    }
-    return templates.TemplateResponse("contract_detail.html", context)
-
-
-def _next_version(ver: str) -> str:
-    v = Version(ver)
-    return f"{v.major}.{v.minor}.{v.micro + 1}"
-
-
-_EXPECTATION_KEYS = (
-    "mustBe",
-    "mustNotBe",
-    "mustBeGreaterThan",
-    "mustBeGreaterOrEqualTo",
-    "mustBeLessThan",
-    "mustBeLessOrEqualTo",
-    "mustBeBetween",
-    "mustNotBeBetween",
-    "query",
 )
 
+template_env = Environment(loader=_env_loader, autoescape=select_autoescape(["html", "xml"]))
+templates = Jinja2Templates(env=template_env)
 
-def _stringify_value(value: Any) -> str:
-    if value is None:
-        return ""
-    if isinstance(value, (bool, int, float)):
-        return json.dumps(value)
-    if isinstance(value, (list, dict)):
-        return json.dumps(value, indent=2, sort_keys=True)
-    return str(value)
+app = FastAPI(title="DC43 Demo Pipeline")
+
+CONTRACTS_APP_URL = os.getenv("DC43_CONTRACTS_APP_URL")
 
 
-def _parse_json_value(raw: Any) -> Any:
-    if raw is None:
-        return None
-    if isinstance(raw, (dict, list, bool, int, float)):
-        return raw
-    if isinstance(raw, str):
-        text = raw.strip()
-        if not text:
-            return None
-        try:
-            return json.loads(text)
-        except json.JSONDecodeError:
-            return text
-    return raw
+@app.get("/")
+async def redirect_to_pipeline() -> RedirectResponse:
+    return RedirectResponse(url="/pipeline-runs", status_code=307)
 
 
-def _as_bool(value: Any) -> Optional[bool]:
-    if value in (None, ""):
-        return None
-    if isinstance(value, bool):
-        return value
-    if isinstance(value, (int, float)):
-        return bool(value)
-    if isinstance(value, str):
-        lowered = value.strip().lower()
-        if lowered in {"true", "1", "yes", "y", "on"}:
-            return True
-        if lowered in {"false", "0", "no", "n", "off"}:
-            return False
-    return bool(value)
+@app.get("/contracts")
+async def redirect_contracts() -> RedirectResponse:
+    if not CONTRACTS_APP_URL:
+        raise HTTPException(status_code=404, detail="Contracts app not configured")
+    return RedirectResponse(url=CONTRACTS_APP_URL, status_code=307)
 
 
-def _as_int(value: Any) -> Optional[int]:
-    if value in (None, ""):
-        return None
-    if isinstance(value, bool):
-        return int(value)
-    try:
-        return int(value)
-    except (TypeError, ValueError):
-        raise ValueError(f"Expected integer value, got {value!r}")
+@app.get("/datasets")
+async def redirect_datasets() -> RedirectResponse:
+    if not CONTRACTS_APP_URL:
+        raise HTTPException(status_code=404, detail="Contracts app not configured")
+    return RedirectResponse(url=f"{CONTRACTS_APP_URL.rstrip('/')}/datasets", status_code=307)
 
 
-def _custom_properties_state(raw: Any) -> List[Dict[str, str]]:
-    state: List[Dict[str, str]] = []
-    for item in normalise_custom_properties(raw):
-        key = None
-        value = None
-        if isinstance(item, Mapping):
-            key = item.get("property")
-            value = item.get("value")
-        else:
-            key = getattr(item, "property", None)
-            value = getattr(item, "value", None)
-        if key:
-            state.append({"property": str(key), "value": _stringify_value(value)})
-    return state
+@app.get("/contracts/{path:path}")
+async def redirect_contract_pages(path: str) -> RedirectResponse:
+    if not CONTRACTS_APP_URL:
+        raise HTTPException(status_code=404, detail="Contracts app not configured")
+    url = f"{CONTRACTS_APP_URL.rstrip('/')}/contracts/{path}" if path else f"{CONTRACTS_APP_URL.rstrip('/')}/contracts"
+    return RedirectResponse(url=url, status_code=307)
 
 
-def _quality_state(items: Optional[Iterable[Any]]) -> List[Dict[str, Any]]:
-    state: List[Dict[str, Any]] = []
-    if not items:
-        return state
-    for item in items:
-        if hasattr(item, "model_dump"):
-            raw = item.model_dump(exclude_none=True)
-        elif hasattr(item, "dict"):
-            raw = item.dict(exclude_none=True)  # type: ignore[attr-defined]
-        else:
-            raw = {k: v for k, v in vars(item).items() if v is not None}
-        expectation = None
-        expectation_value = None
-        for key in _EXPECTATION_KEYS:
-            if key in raw:
-                expectation = key
-                expectation_value = raw.pop(key)
-                break
-        for key, value in list(raw.items()):
-            if isinstance(value, (list, dict)):
-                raw[key] = json.dumps(value, indent=2, sort_keys=True)
-        entry: Dict[str, Any] = {k: v for k, v in raw.items() if v is not None}
-        if expectation:
-            entry["expectation"] = expectation
-            if isinstance(expectation_value, list):
-                entry["expectationValue"] = ", ".join(str(v) for v in expectation_value)
-            elif isinstance(expectation_value, (dict, list)):
-                entry["expectationValue"] = json.dumps(expectation_value, indent=2, sort_keys=True)
-            elif expectation_value is None:
-                entry["expectationValue"] = ""
-            else:
-                entry["expectationValue"] = str(expectation_value)
-        state.append(entry)
-    return state
-
-
-def _schema_property_state(prop: SchemaProperty) -> Dict[str, Any]:
-    examples = getattr(prop, "examples", None) or []
-    return {
-        "name": getattr(prop, "name", "") or "",
-        "physicalType": getattr(prop, "physicalType", "") or "",
-        "description": getattr(prop, "description", "") or "",
-        "businessName": getattr(prop, "businessName", "") or "",
-        "logicalType": getattr(prop, "logicalType", "") or "",
-        "logicalTypeOptions": _stringify_value(getattr(prop, "logicalTypeOptions", None)),
-        "required": bool(getattr(prop, "required", False)),
-        "unique": bool(getattr(prop, "unique", False)),
-        "partitioned": bool(getattr(prop, "partitioned", False)),
-        "primaryKey": bool(getattr(prop, "primaryKey", False)),
-        "classification": getattr(prop, "classification", "") or "",
-        "examples": "\n".join(str(item) for item in examples),
-        "customProperties": _custom_properties_state(getattr(prop, "customProperties", None)),
-        "quality": _quality_state(getattr(prop, "quality", None)),
-    }
-
-
-def _schema_object_state(obj: SchemaObject) -> Dict[str, Any]:
-    properties = [
-        _schema_property_state(prop)
-        for prop in getattr(obj, "properties", None) or []
-    ]
-    return {
-        "name": getattr(obj, "name", "") or "",
-        "description": getattr(obj, "description", "") or "",
-        "businessName": getattr(obj, "businessName", "") or "",
-        "logicalType": getattr(obj, "logicalType", "") or "",
-        "customProperties": _custom_properties_state(getattr(obj, "customProperties", None)),
-        "quality": _quality_state(getattr(obj, "quality", None)),
-        "properties": properties,
-    }
-
-
-_SERVER_FIELD_MAP = {
-    "description": "description",
-    "environment": "environment",
-    "format": "format",
-    "path": "path",
-    "dataset": "dataset",
-    "database": "database",
-    "schema": "schema_",
-    "catalog": "catalog",
-    "host": "host",
-    "location": "location",
-    "endpointUrl": "endpointUrl",
-    "project": "project",
-    "region": "region",
-    "regionName": "regionName",
-    "serviceName": "serviceName",
-    "warehouse": "warehouse",
-    "stagingDir": "stagingDir",
-    "account": "account",
-}
-
-
-def _server_state(server: Server) -> Dict[str, Any]:
-    state = {
-        "server": getattr(server, "server", "") or "",
-        "type": getattr(server, "type", "") or "",
-        "port": getattr(server, "port", None) or "",
-    }
-    for field, attr in _SERVER_FIELD_MAP.items():
-        state[field] = getattr(server, attr, "") or ""
-    versioning_value: Any | None = None
-    path_pattern_value: Any | None = None
-    custom_entries: List[Dict[str, str]] = []
-    for item in normalise_custom_properties(getattr(server, "customProperties", None)):
-        key = None
-        value = None
-        if isinstance(item, Mapping):
-            key = item.get("property")
-            value = item.get("value")
-        else:
-            key = getattr(item, "property", None)
-            value = getattr(item, "value", None)
-        if not key:
-            continue
-        if str(key) == "dc43.versioning":
-            versioning_value = value
-            continue
-        if str(key) == "dc43.pathPattern":
-            path_pattern_value = value
-            continue
-        custom_entries.append({"property": str(key), "value": _stringify_value(value)})
-    if versioning_value is not None:
-        parsed = versioning_value
-        if isinstance(parsed, str):
-            parsed = _parse_json_value(parsed)
-        state["versioningConfig"] = parsed if isinstance(parsed, Mapping) else None
-    if path_pattern_value not in (None, ""):
-        state["pathPattern"] = str(path_pattern_value)
-    state["customProperties"] = custom_entries
-    return state
-
-
-def _support_state(items: Optional[Iterable[Support]]) -> List[Dict[str, Any]]:
-    result: List[Dict[str, Any]] = []
-    if not items:
-        return result
-    for entry in items:
-        payload: Dict[str, Any] = {}
-        for field in ("channel", "url", "description", "tool", "scope", "invitationUrl"):
-            value = getattr(entry, field, None)
-            if value:
-                payload[field] = value
-        if payload:
-            result.append(payload)
-    return result
-
-
-def _sla_state(items: Optional[Iterable[ServiceLevelAgreementProperty]]) -> List[Dict[str, Any]]:
-    result: List[Dict[str, Any]] = []
-    if not items:
-        return result
-    for entry in items:
-        payload: Dict[str, Any] = {}
-        for field in ("property", "value", "valueExt", "unit", "element", "driver"):
-            value = getattr(entry, field, None)
-            if value is None:
-                continue
-            if field in {"value", "valueExt"}:
-                payload[field] = _stringify_value(value)
-            else:
-                payload[field] = value
-        if payload:
-            result.append(payload)
-    return result
-
-
-def _contract_editor_state(contract: Optional[OpenDataContractStandard] = None) -> Dict[str, Any]:
-    if contract is None:
-        return {
-            "id": "",
-            "version": "",
-            "kind": "DataContract",
-            "apiVersion": "3.0.2",
-            "name": "",
-            "description": "",
-            "status": "",
-            "domain": "",
-            "dataProduct": "",
-            "tenant": "",
-            "tags": [],
-            "customProperties": [],
-            "servers": [],
-            "schemaObjects": [
-                {
-                    "name": "",
-                    "description": "",
-                    "businessName": "",
-                    "logicalType": "",
-                    "customProperties": [],
-                    "quality": [],
-                    "properties": [],
-                }
-            ],
-            "support": [],
-            "slaProperties": [],
-        }
-    description = getattr(contract.description, "usage", "") if getattr(contract, "description", None) else ""
-    state = {
-        "id": getattr(contract, "id", "") or "",
-        "version": getattr(contract, "version", "") or "",
-        "kind": getattr(contract, "kind", "DataContract") or "DataContract",
-        "apiVersion": getattr(contract, "apiVersion", "3.0.2") or "3.0.2",
-        "name": getattr(contract, "name", "") or "",
-        "description": description,
-        "status": getattr(contract, "status", "") or "",
-        "domain": getattr(contract, "domain", "") or "",
-        "dataProduct": getattr(contract, "dataProduct", "") or "",
-        "tenant": getattr(contract, "tenant", "") or "",
-        "tags": list(getattr(contract, "tags", []) or []),
-        "customProperties": _custom_properties_state(getattr(contract, "customProperties", None)),
-        "servers": [_server_state(server) for server in getattr(contract, "servers", []) or []],
-        "schemaObjects": [
-            _schema_object_state(obj) for obj in getattr(contract, "schema_", None) or []
-        ],
-        "support": _support_state(getattr(contract, "support", None)),
-        "slaProperties": _sla_state(getattr(contract, "slaProperties", None)),
-    }
-    if not state["schemaObjects"]:
-        state["schemaObjects"] = _contract_editor_state(None)["schemaObjects"]
-    return state
-
-
-def _sorted_versions(values: Iterable[str]) -> List[str]:
-    parsed: List[Tuple[Version, str]] = []
-    invalid: List[str] = []
-    for value in values:
-        if not value:
-            continue
-        try:
-            parsed.append((Version(str(value)), str(value)))
-        except InvalidVersion:
-            invalid.append(str(value))
-    parsed.sort(key=lambda entry: entry[0])
-    return [ver for _, ver in parsed] + sorted(invalid)
-
-
-def _build_editor_meta(
-    *,
-    editor_state: Mapping[str, Any],
-    editing: bool,
-    original_version: Optional[str],
-    baseline_state: Optional[Mapping[str, Any]],
-    baseline_contract: Optional[OpenDataContractStandard],
-) -> Dict[str, Any]:
-    existing_contracts = sorted(store.list_contracts())
-    version_map: Dict[str, List[str]] = {}
-    for contract_id in existing_contracts:
-        try:
-            versions = store.list_versions(contract_id)
-        except FileNotFoundError:
-            versions = []
-        version_map[contract_id] = _sorted_versions(versions)
-    meta: Dict[str, Any] = {
-        "existingContracts": existing_contracts,
-        "existingVersions": version_map,
-        "editing": editing,
-        "originalVersion": original_version,
-        "contractId": str(editor_state.get("id", "")) or (
-            getattr(baseline_contract, "id", "") if baseline_contract else ""
-        ),
-    }
-    if original_version:
-        meta["baseVersion"] = original_version
-    if baseline_state is None and baseline_contract is not None:
-        baseline_state = _contract_editor_state(baseline_contract)
-    if baseline_state is not None:
-        # ensure baseline is JSON serializable
-        meta["baselineState"] = jsonable_encoder(baseline_state)
-    if baseline_contract is not None:
-        meta["baseContract"] = contract_to_dict(baseline_contract)
-    return meta
-
-
-def _editor_context(
-    request: Request,
-    *,
-    editor_state: Dict[str, Any],
-    editing: bool = False,
-    original_version: Optional[str] = None,
-    baseline_state: Optional[Mapping[str, Any]] = None,
-    baseline_contract: Optional[OpenDataContractStandard] = None,
-    error: Optional[str] = None,
-) -> Dict[str, Any]:
-    context = {
-        "request": request,
-        "editing": editing,
-        "editor_state": editor_state,
-        "status_options": _STATUS_OPTIONS,
-        "versioning_modes": _VERSIONING_MODES,
-        "editor_meta": _build_editor_meta(
-            editor_state=editor_state,
-            editing=editing,
-            original_version=original_version,
-            baseline_state=baseline_state,
-            baseline_contract=baseline_contract,
-        ),
-    }
-    if original_version:
-        context["original_version"] = original_version
-    if error:
-        context["error"] = error
-    return context
-
-
-def _custom_properties_models(items: Optional[Iterable[Mapping[str, Any]]]) -> List[CustomProperty] | None:
-    result: List[CustomProperty] = []
-    if not items:
-        return None
-    for item in items:
-        if not isinstance(item, Mapping):
-            continue
-        key = (str(item.get("property", ""))).strip()
-        if not key:
-            continue
-        value = _parse_json_value(item.get("value"))
-        result.append(CustomProperty(property=key, value=value))
-    return result or None
-
-
-def _validate_contract_payload(
-    payload: Mapping[str, Any],
-    *,
-    editing: bool,
-    base_contract_id: Optional[str] = None,
-    base_version: Optional[str] = None,
-) -> None:
-    contract_id = (str(payload.get("id", ""))).strip()
-    if not contract_id:
-        raise ValueError("Contract ID is required")
-    version = (str(payload.get("version", ""))).strip()
-    if not version:
-        raise ValueError("Version is required")
-    try:
-        new_version = SemVer.parse(version)
-    except ValueError as exc:
-        raise ValueError(f"Invalid semantic version: {exc}") from exc
-    existing_contracts = set(store.list_contracts())
-    existing_versions = (
-        set(store.list_versions(contract_id)) if contract_id in existing_contracts else set()
-    )
-    if editing:
-        if base_contract_id and contract_id != base_contract_id:
-            raise ValueError("Contract ID cannot be changed while editing")
-        if base_version:
-            try:
-                prior = SemVer.parse(base_version)
-            except ValueError:
-                prior = None
-            if prior and (
-                (new_version.major, new_version.minor, new_version.patch)
-                <= (prior.major, prior.minor, prior.patch)
-            ):
-                raise ValueError(
-                    f"Version {version} must be greater than {base_version}"
-                )
-        if version in existing_versions:
-            raise ValueError(
-                f"Version {version} is already stored for contract {contract_id}"
-            )
-    else:
-        if contract_id in existing_contracts and version in existing_versions:
-            raise ValueError(
-                f"Contract {contract_id} already has a version {version}."
-            )
-
-
-def _support_models(items: Optional[Iterable[Mapping[str, Any]]]) -> List[Support] | None:
-    result: List[Support] = []
-    if not items:
-        return None
-    for item in items:
-        if not isinstance(item, Mapping):
-            continue
-        channel = (str(item.get("channel", ""))).strip()
-        if not channel:
-            continue
-        payload: Dict[str, Any] = {"channel": channel}
-        for field in ("url", "description", "tool", "scope", "invitationUrl"):
-            value = item.get(field)
-            if value:
-                payload[field] = value
-        result.append(Support(**payload))
-    return result or None
-
-
-def _sla_models(items: Optional[Iterable[Mapping[str, Any]]]) -> List[ServiceLevelAgreementProperty] | None:
-    result: List[ServiceLevelAgreementProperty] = []
-    if not items:
-        return None
-    for item in items:
-        if not isinstance(item, Mapping):
-            continue
-        key = (str(item.get("property", ""))).strip()
-        if not key:
-            continue
-        payload: Dict[str, Any] = {"property": key}
-        for field in ("unit", "element", "driver"):
-            value = item.get(field)
-            if value:
-                payload[field] = value
-        value = item.get("value")
-        if value not in (None, ""):
-            payload["value"] = _parse_json_value(value)
-        value_ext = item.get("valueExt")
-        if value_ext not in (None, ""):
-            payload["valueExt"] = _parse_json_value(value_ext)
-        result.append(ServiceLevelAgreementProperty(**payload))
-    return result or None
-
-
-def _parse_expectation_value(expectation: str, value: Any) -> Any:
-    if value is None or value == "":
-        return None
-    if isinstance(value, (list, dict, bool, int, float)):
-        return value
-    if not isinstance(value, str):
-        return value
-    text = value.strip()
-    if expectation in {"mustBeBetween", "mustNotBeBetween"}:
-        separators = [",", ";"]
-        for sep in separators:
-            if sep in text:
-                parts = [p.strip() for p in text.split(sep) if p.strip()]
-                break
-        else:
-            parts = [p.strip() for p in text.split() if p.strip()]
-        if len(parts) < 2:
-            raise ValueError("Data quality range requires two numeric values")
-        try:
-            return [float(parts[0]), float(parts[1])]
-        except ValueError as exc:
-            raise ValueError("Data quality range must be numeric") from exc
-    if expectation in {
-        "mustBeGreaterThan",
-        "mustBeGreaterOrEqualTo",
-        "mustBeLessThan",
-        "mustBeLessOrEqualTo",
-    }:
-        try:
-            return float(text)
-        except ValueError as exc:
-            raise ValueError(f"Expectation {expectation} requires a numeric value") from exc
-    if expectation in {"mustBe", "mustNotBe"}:
-        try:
-            return json.loads(text)
-        except json.JSONDecodeError:
-            return text
-    return text
-
-
-def _quality_models(items: Optional[Iterable[Mapping[str, Any]]]) -> List[DataQuality] | None:
-    result: List[DataQuality] = []
-    if not items:
-        return None
-    for item in items:
-        if not isinstance(item, Mapping):
-            continue
-        payload: Dict[str, Any] = {}
-        for field in (
-            "name",
-            "type",
-            "rule",
-            "description",
-            "dimension",
-            "severity",
-            "unit",
-            "schedule",
-            "scheduler",
-            "businessImpact",
-            "method",
-        ):
-            value = item.get(field)
-            if value not in (None, ""):
-                payload[field] = value
-        tags_value = item.get("tags")
-        if isinstance(tags_value, str):
-            tags = [t.strip() for t in tags_value.split(",") if t.strip()]
-            if tags:
-                payload["tags"] = tags
-        elif isinstance(tags_value, Iterable):
-            tags = [str(t).strip() for t in tags_value if str(t).strip()]
-            if tags:
-                payload["tags"] = tags
-        expectation = item.get("expectation")
-        if expectation:
-            payload[expectation] = _parse_expectation_value(expectation, item.get("expectationValue"))
-        implementation = item.get("implementation")
-        if implementation not in (None, ""):
-            payload["implementation"] = _parse_json_value(implementation)
-        custom_props = _custom_properties_models(item.get("customProperties"))
-        if custom_props:
-            payload["customProperties"] = custom_props
-        if payload:
-            result.append(DataQuality(**payload))
-    return result or None
-
-
-def _schema_properties_models(items: Optional[Iterable[Mapping[str, Any]]]) -> List[SchemaProperty]:
-    result: List[SchemaProperty] = []
-    if not items:
-        return result
-    for item in items:
-        if not isinstance(item, Mapping):
-            continue
-        name = (str(item.get("name", ""))).strip()
-        if not name:
-            continue
-        payload: Dict[str, Any] = {"name": name}
-        physical_type = item.get("physicalType")
-        if physical_type:
-            payload["physicalType"] = physical_type
-        for field in (
-            "description",
-            "businessName",
-            "classification",
-            "logicalType",
-        ):
-            value = item.get(field)
-            if value not in (None, ""):
-                payload[field] = value
-        logical_type_options = item.get("logicalTypeOptions")
-        if logical_type_options not in (None, ""):
-            payload["logicalTypeOptions"] = _parse_json_value(logical_type_options)
-        for boolean_field in ("required", "unique", "partitioned", "primaryKey"):
-            value = _as_bool(item.get(boolean_field))
-            if value is not None:
-                payload[boolean_field] = value
-        examples = item.get("examples")
-        if isinstance(examples, str):
-            values = [ex.strip() for ex in examples.splitlines() if ex.strip()]
-            if values:
-                payload["examples"] = values
-        elif isinstance(examples, Iterable):
-            values = [str(ex).strip() for ex in examples if str(ex).strip()]
-            if values:
-                payload["examples"] = values
-        custom_props = _custom_properties_models(item.get("customProperties"))
-        if custom_props:
-            payload["customProperties"] = custom_props
-        quality = _quality_models(item.get("quality"))
-        if quality:
-            payload["quality"] = quality
-        result.append(SchemaProperty(**payload))
-    return result
-
-
-def _schema_objects_models(items: Optional[Iterable[Mapping[str, Any]]]) -> List[SchemaObject]:
-    result: List[SchemaObject] = []
-    if not items:
-        return result
-    for item in items:
-        if not isinstance(item, Mapping):
-            continue
-        name = (str(item.get("name", ""))).strip()
-        payload: Dict[str, Any] = {}
-        if name:
-            payload["name"] = name
-        for field in ("description", "businessName", "logicalType"):
-            value = item.get(field)
-            if value not in (None, ""):
-                payload[field] = value
-        custom_props = _custom_properties_models(item.get("customProperties"))
-        if custom_props:
-            payload["customProperties"] = custom_props
-        quality = _quality_models(item.get("quality"))
-        if quality:
-            payload["quality"] = quality
-        properties = _schema_properties_models(item.get("properties"))
-        if properties:
-            name_counts = Counter(
-                prop.name for prop in properties if getattr(prop, "name", None)
-            )
-            duplicates = [name for name, count in name_counts.items() if count > 1]
-            if duplicates:
-                object_name = payload.get("name") or "schema object"
-                dup_list = ", ".join(sorted(duplicates))
-                raise ValueError(
-                    f"Duplicate field name(s) {dup_list} in {object_name}"
-                )
-        payload["properties"] = properties
-        result.append(SchemaObject(**payload))
-    return result
-
-
-def _server_models(items: Optional[Iterable[Mapping[str, Any]]]) -> List[Server] | None:
-    result: List[Server] = []
-    if not items:
-        return None
-    for item in items:
-        if not isinstance(item, Mapping):
-            continue
-        server_name = (str(item.get("server", ""))).strip()
-        server_type = (str(item.get("type", ""))).strip()
-        if not server_name or not server_type:
-            continue
-        payload: Dict[str, Any] = {"server": server_name, "type": server_type}
-        for field, attr in _SERVER_FIELD_MAP.items():
-            value = item.get(field)
-            if value not in (None, ""):
-                payload[attr] = value
-        port_value = item.get("port")
-        if port_value not in (None, ""):
-            payload["port"] = _as_int(port_value)
-        custom_props: List[CustomProperty] = []
-        base_custom = _custom_properties_models(item.get("customProperties"))
-        if base_custom:
-            custom_props.extend(base_custom)
-        versioning_config = item.get("versioningConfig")
-        if versioning_config not in (None, "", {}):
-            parsed_versioning = (
-                versioning_config
-                if isinstance(versioning_config, Mapping)
-                else _parse_json_value(versioning_config)
-            )
-            if not isinstance(parsed_versioning, Mapping):
-                raise ValueError("dc43.versioning must be provided as an object")
-            custom_props.append(
-                CustomProperty(property="dc43.versioning", value=parsed_versioning)
-            )
-        path_pattern = item.get("pathPattern")
-        if path_pattern not in (None, ""):
-            custom_props.append(
-                CustomProperty(property="dc43.pathPattern", value=str(path_pattern))
-            )
-        if custom_props:
-            payload["customProperties"] = custom_props
-        result.append(Server(**payload))
-    return result or None
-
-
-def _normalise_tags(value: Any) -> List[str] | None:
-    if value in (None, ""):
-        return None
-    if isinstance(value, str):
-        tags = [item.strip() for item in value.split(",") if item.strip()]
-        return tags or None
-    if isinstance(value, Iterable):
-        tags = [str(item).strip() for item in value if str(item).strip()]
-        return tags or None
-    return None
-
-
-def _build_contract_from_payload(payload: Mapping[str, Any]) -> OpenDataContractStandard:
-    contract_id = (str(payload.get("id", ""))).strip()
-    if not contract_id:
-        raise ValueError("Contract ID is required")
-    version = (str(payload.get("version", ""))).strip()
-    if not version:
-        raise ValueError("Version is required")
-    name = (str(payload.get("name", "")) or contract_id).strip()
-    description = str(payload.get("description", ""))
-    kind = (str(payload.get("kind", "DataContract")) or "DataContract").strip()
-    api_version = (str(payload.get("apiVersion", "3.0.2")) or "3.0.2").strip()
-    status = str(payload.get("status", "")).strip() or None
-    domain = str(payload.get("domain", "")).strip() or None
-    data_product = str(payload.get("dataProduct", "")).strip() or None
-    tenant = str(payload.get("tenant", "")).strip() or None
-    tags = _normalise_tags(payload.get("tags"))
-    custom_props = _custom_properties_models(payload.get("customProperties"))
-    servers = _server_models(payload.get("servers"))
-    schema_objects = _schema_objects_models(payload.get("schemaObjects"))
-    if not schema_objects:
-        raise ValueError("At least one schema object with fields is required")
-    # Ensure each schema object has properties
-    for obj in schema_objects:
-        if not obj.properties:
-            raise ValueError("Each schema object must define at least one field")
-    support_entries = _support_models(payload.get("support"))
-    sla_properties = _sla_models(payload.get("slaProperties"))
-    return OpenDataContractStandard(
-        version=version,
-        kind=kind,
-        apiVersion=api_version,
-        id=contract_id,
-        name=name,
-        description=None if not description else Description(usage=description),
-        status=status,
-        domain=domain,
-        dataProduct=data_product,
-        tenant=tenant,
-        tags=tags,
-        customProperties=custom_props,
-        servers=servers,
-        schema=schema_objects,  # type: ignore[arg-type]
-        support=support_entries,
-        slaProperties=sla_properties,
-    )
-
-
-@app.get("/contracts/{cid}/{ver}/edit", response_class=HTMLResponse)
-async def edit_contract_form(request: Request, cid: str, ver: str) -> HTMLResponse:
-    try:
-        contract = store.get(cid, ver)
-    except FileNotFoundError as exc:
-        raise HTTPException(status_code=404, detail=str(exc))
-    new_ver = _next_version(ver)
-    editor_state = _contract_editor_state(contract)
-    baseline_state = json.loads(json.dumps(editor_state))
-    editor_state["version"] = new_ver
-    context = _editor_context(
-        request,
-        editor_state=editor_state,
-        editing=True,
-        original_version=ver,
-        baseline_state=baseline_state,
-        baseline_contract=contract,
-    )
-    return templates.TemplateResponse("new_contract.html", context)
-
-
-@app.post("/contracts/{cid}/{ver}/edit", response_class=HTMLResponse)
-async def save_contract_edits(
-    request: Request,
-    cid: str,
-    ver: str,
-    payload: str = Form(...),
-    original_version: str = Form(""),
-) -> HTMLResponse:
-    editor_state: Dict[str, Any]
-    baseline_contract: Optional[OpenDataContractStandard] = None
-    baseline_state: Optional[Dict[str, Any]] = None
-    base_version = original_version or ver
-    try:
-        baseline_contract = store.get(cid, base_version)
-        baseline_state = json.loads(json.dumps(_contract_editor_state(baseline_contract)))
-    except FileNotFoundError:
-        baseline_contract = None
-        baseline_state = None
-    try:
-        editor_state = json.loads(payload)
-    except json.JSONDecodeError as exc:
-        error = f"Invalid editor payload: {exc.msg}"
-        editor_state = _contract_editor_state()
-        editor_state["id"] = cid
-        editor_state["version"] = _next_version(ver)
-    else:
-        try:
-            _validate_contract_payload(
-                editor_state,
-                editing=True,
-                base_contract_id=cid,
-                base_version=base_version,
-            )
-            model = _build_contract_from_payload(editor_state)
-            store.put(model)
-            return RedirectResponse(
-                url=f"/contracts/{model.id}/{model.version}", status_code=303
-            )
-        except (ValidationError, ValueError) as exc:
-            error = str(exc)
-        except Exception as exc:  # pragma: no cover - display unexpected errors
-            error = str(exc)
-    context = _editor_context(
-        request,
-        editor_state=editor_state,
-        editing=True,
-        original_version=base_version,
-        baseline_state=baseline_state,
-        baseline_contract=baseline_contract,
-        error=error,
-    )
-    return templates.TemplateResponse("new_contract.html", context)
-
-
-@app.post("/contracts/{cid}/{ver}/validate")
-async def html_validate_contract(cid: str, ver: str) -> HTMLResponse:
-    return RedirectResponse(url=f"/contracts/{cid}/{ver}", status_code=303)
-
-
-@app.get("/datasets", response_class=HTMLResponse)
-async def list_datasets(request: Request) -> HTMLResponse:
-    catalog = dataset_catalog(load_records())
-    context = {"request": request, "datasets": catalog}
-    return templates.TemplateResponse("datasets.html", context)
+@app.get("/datasets/{path:path}")
+async def redirect_dataset_pages(path: str) -> RedirectResponse:
+    if not CONTRACTS_APP_URL:
+        raise HTTPException(status_code=404, detail="Contracts app not configured")
+    url = f"{CONTRACTS_APP_URL.rstrip('/')}/datasets/{path}" if path else f"{CONTRACTS_APP_URL.rstrip('/')}/datasets"
+    return RedirectResponse(url=url, status_code=307)
 
 
 @app.get("/pipeline-runs", response_class=HTMLResponse)
@@ -2712,178 +104,6 @@ async def list_pipeline_runs(request: Request) -> HTMLResponse:
     return templates.TemplateResponse("pipeline_runs.html", context)
 
 
-@app.get("/datasets/{dataset_name}", response_class=HTMLResponse)
-async def dataset_versions(request: Request, dataset_name: str) -> HTMLResponse:
-    records = [r.__dict__.copy() for r in load_records() if r.dataset_name == dataset_name]
-    context = {"request": request, "dataset_name": dataset_name, "records": records}
-    return templates.TemplateResponse("dataset_versions.html", context)
-
-
-def _dataset_path(contract: OpenDataContractStandard | None, dataset_name: str, dataset_version: str) -> Path:
-    server = (contract.servers or [None])[0] if contract else None
-    data_root = Path(DATA_DIR).parent
-    base = Path(getattr(server, "path", "")) if server else data_root
-    if base.suffix:
-        base = base.parent
-    if not base.is_absolute():
-        base = data_root / base
-    if base.name == dataset_name:
-        return base / dataset_version
-    return base / dataset_name / dataset_version
-
-
-def _dataset_preview(contract: OpenDataContractStandard | None, dataset_name: str, dataset_version: str) -> str:
-    ds_path = _dataset_path(contract, dataset_name, dataset_version)
-    server = (contract.servers or [None])[0] if contract else None
-    fmt = getattr(server, "format", None)
-    try:
-        if fmt == "parquet":
-            from pyspark.sql import SparkSession  # type: ignore
-            spark = SparkSession.builder.master("local[1]").appName("preview").getOrCreate()
-            df = spark.read.parquet(str(ds_path))
-            return "\n".join(str(r.asDict()) for r in df.limit(10).collect())[:1000]
-        if fmt == "json":
-            target = ds_path if ds_path.is_file() else next(ds_path.glob("*.json"), None)
-            if target:
-                return target.read_text()[:1000]
-        if ds_path.is_file():
-            return ds_path.read_text()[:1000]
-        if ds_path.is_dir():
-            target = next((p for p in ds_path.iterdir() if p.is_file()), None)
-            if target:
-                return target.read_text()[:1000]
-    except Exception:
-        return ""
-    return ""
-
-
-def dataset_catalog(records: Iterable[DatasetRecord]) -> List[Dict[str, Any]]:
-    """Summarise known datasets and associated contract information."""
-
-    grouped: Dict[str, Dict[str, Any]] = {}
-    for record in records:
-        if not record.dataset_name:
-            continue
-        bucket = grouped.setdefault(
-            record.dataset_name,
-            {"dataset_name": record.dataset_name, "records": []},
-        )
-        bucket["records"].append(record)
-
-    for contract_id in store.list_contracts():
-        for version in store.list_versions(contract_id):
-            try:
-                contract = store.get(contract_id, version)
-            except FileNotFoundError:
-                continue
-            server_info = _server_details(contract) or {}
-            dataset_id = (
-                server_info.get("dataset_id")
-                or server_info.get("dataset")
-                or contract.id
-                or contract_id
-            )
-            bucket = grouped.setdefault(
-                dataset_id,
-                {"dataset_name": dataset_id, "records": []},
-            )
-            contracts_map = bucket.setdefault("contracts_by_id", {})
-            contracts = contracts_map.setdefault(contract_id, [])
-            contracts.append(
-                {
-                    "version": version,
-                    "status": getattr(contract, "status", ""),
-                    "server": server_info,
-                }
-            )
-
-    catalog: List[Dict[str, Any]] = []
-    for dataset_name, payload in grouped.items():
-        dataset_records: List[DatasetRecord] = list(payload.get("records", []))
-        dataset_records.sort(key=lambda item: _version_sort_key(item.dataset_version or ""))
-        latest_record: Optional[DatasetRecord] = dataset_records[-1] if dataset_records else None
-
-        latest_status_value: Optional[str] = None
-        latest_status_label: str = ""
-        latest_status_badge: str = ""
-        latest_reason: Optional[str] = None
-        if latest_record:
-            status_raw = str(latest_record.status or "unknown")
-            latest_status_value = status_raw.lower()
-            latest_status_label = status_raw.replace("_", " ").title()
-            latest_status_badge = _DQ_STATUS_BADGES.get(latest_status_value, "bg-secondary")
-            latest_reason = latest_record.reason or None
-
-        run_drafts = sorted(
-            {rec.draft_contract_version for rec in dataset_records if rec.draft_contract_version},
-            key=_version_sort_key,
-        )
-        contracts_summary: List[Dict[str, Any]] = []
-        contracts_map: Dict[str, List[Dict[str, Any]]] = payload.get("contracts_by_id", {})
-        for contract_id, versions in contracts_map.items():
-            versions.sort(key=lambda item: _version_sort_key(item["version"]))
-            other_versions = [item["version"] for item in versions[:-1]]
-            latest_contract = versions[-1] if versions else {"version": "", "status": ""}
-            status_raw = str(latest_contract.get("status") or "unknown")
-            status_value = status_raw.lower()
-            status_label = status_raw.replace("_", " ").title()
-            draft_versions = [
-                item for item in versions if str(item.get("status", "")).lower() == "draft"
-            ]
-            latest_draft = draft_versions[-1]["version"] if draft_versions else None
-            contracts_summary.append(
-                {
-                    "id": contract_id,
-                    "latest_version": latest_contract.get("version", ""),
-                    "latest_status": status_value,
-                    "latest_status_label": status_label,
-                    "other_versions": other_versions,
-                    "drafts_count": len(draft_versions),
-                    "latest_draft_version": latest_draft,
-                }
-            )
-
-        contracts_summary.sort(key=lambda item: item["id"])
-
-        catalog.append(
-            {
-                "dataset_name": dataset_name,
-                "latest_version": latest_record.dataset_version if latest_record else "",
-                "latest_status": latest_status_value,
-                "latest_status_label": latest_status_label,
-                "latest_status_badge": latest_status_badge,
-                "latest_record_reason": latest_reason,
-                "contract_summaries": contracts_summary,
-                "run_drafts_count": len(run_drafts),
-                "run_latest_draft_version": run_drafts[-1] if run_drafts else None,
-            }
-        )
-
-    catalog.sort(key=lambda item: item["dataset_name"])
-    return catalog
-
-
-@app.get("/datasets/{dataset_name}/{dataset_version}", response_class=HTMLResponse)
-async def dataset_detail(request: Request, dataset_name: str, dataset_version: str) -> HTMLResponse:
-    for r in load_records():
-        if r.dataset_name == dataset_name and r.dataset_version == dataset_version:
-            contract_obj: OpenDataContractStandard | None = None
-            if r.contract_id and r.contract_version:
-                try:
-                    contract_obj = store.get(r.contract_id, r.contract_version)
-                except FileNotFoundError:
-                    contract_obj = None
-            preview = _dataset_preview(contract_obj, dataset_name, dataset_version)
-            context = {
-                "request": request,
-                "record": r,
-                "contract": contract_to_dict(contract_obj) if contract_obj else None,
-                "data_preview": preview,
-            }
-            return templates.TemplateResponse("dataset_detail.html", context)
-    raise HTTPException(status_code=404, detail="Dataset not found")
-
-
 @app.post("/pipeline/run", response_class=HTMLResponse)
 async def run_pipeline_endpoint(scenario: str = Form(...)) -> HTMLResponse:
     from .pipeline import run_pipeline
@@ -2892,7 +112,7 @@ async def run_pipeline_endpoint(scenario: str = Form(...)) -> HTMLResponse:
     if not cfg:
         params = urlencode({"error": f"Unknown scenario: {scenario}"})
         return RedirectResponse(url=f"/datasets?{params}", status_code=303)
-    p = cfg["params"]
+    params_cfg = cfg["params"]
     for dataset, version in cfg.get("activate_versions", {}).items():
         try:
             set_active_version(dataset, version)
@@ -2901,67 +121,28 @@ async def run_pipeline_endpoint(scenario: str = Form(...)) -> HTMLResponse:
     try:
         dataset_name, new_version = await asyncio.to_thread(
             run_pipeline,
-            p.get("contract_id"),
-            p.get("contract_version"),
-            p.get("dataset_name"),
-            p.get("dataset_version"),
-            p.get("run_type", "infer"),
-            p.get("collect_examples", False),
-            p.get("examples_limit", 5),
-            p.get("violation_strategy"),
-            p.get("inputs"),
-            p.get("output_adjustment"),
+            params_cfg.get("contract_id"),
+            params_cfg.get("contract_version"),
+            params_cfg.get("dataset_name"),
+            params_cfg.get("dataset_version"),
+            params_cfg.get("run_type", "infer"),
+            params_cfg.get("collect_examples", False),
+            params_cfg.get("examples_limit", 5),
+            params_cfg.get("violation_strategy"),
+            params_cfg.get("inputs"),
+            params_cfg.get("output_adjustment"),
             scenario_key=scenario,
         )
-        label = dataset_name or p.get("dataset_name") or p.get("contract_id") or "dataset"
+        label = (
+            dataset_name
+            or params_cfg.get("dataset_name")
+            or params_cfg.get("contract_id")
+            or "dataset"
+        )
         token = queue_flash(message=f"Run succeeded: {label} {new_version}")
-        params = urlencode({"flash": token})
+        params_qs = urlencode({"flash": token})
     except Exception as exc:  # pragma: no cover - surface pipeline errors
         logger.exception("Pipeline run failed for scenario %s", scenario)
         token = queue_flash(error=str(exc))
-        params = urlencode({"flash": token})
-    return RedirectResponse(url=f"/pipeline-runs?{params}", status_code=303)
-
-
-def run() -> None:  # pragma: no cover - convenience runner
-    """Run the demo UI and spawn a dedicated backend server."""
-
-    import uvicorn
-
-    backend_host = os.getenv("DC43_DEMO_BACKEND_HOST", "127.0.0.1")
-    backend_port = int(os.getenv("DC43_DEMO_BACKEND_PORT", "8001"))
-    backend_url = f"http://{backend_host}:{backend_port}"
-
-    env = os.environ.copy()
-    env.setdefault("DC43_CONTRACT_STORE", str(CONTRACT_DIR))
-    cmd = [
-        sys.executable,
-        "-m",
-        "uvicorn",
-        "dc43_service_backends.webapp:app",
-        "--host",
-        backend_host,
-        "--port",
-        str(backend_port),
-    ]
-    log_level = os.getenv("DC43_DEMO_BACKEND_LOG")
-    if log_level:
-        cmd.extend(["--log-level", log_level])
-
-    process = subprocess.Popen(cmd, env=env)
-
-    try:
-        _wait_for_backend(backend_url)
-    except Exception:
-        process.terminate()
-        process.wait(timeout=5)
-        raise
-
-    try:
-        _initialise_backend(base_url=backend_url)
-        uvicorn.run("dc43.demo_app.server:app", host="0.0.0.0", port=8000)
-    finally:
-        process.terminate()
-        with contextlib.suppress(Exception):
-            process.wait(timeout=5)
-        _initialise_backend(base_url=os.getenv("DC43_DEMO_BACKEND_URL"))
+        params_qs = urlencode({"flash": token})
+    return RedirectResponse(url=f"/pipeline-runs?{params_qs}", status_code=303)
