+++ tests/test_server_pages.py
@@ -1,17 +1,20 @@
 import pytest
 from fastapi.testclient import TestClient
 
-from dc43.demo_app.server import (
-    SCENARIOS,
-    DatasetRecord,
-    _dq_version_records,
-    app,
-    load_records,
-    queue_flash,
-    save_records,
-    scenario_run_rows,
-    store,
-)
+from dc43.contracts_app import server as contracts_server
+from dc43.demo_app import server as demo_server
+
+SCENARIOS = contracts_server.SCENARIOS
+DatasetRecord = contracts_server.DatasetRecord
+_dq_version_records = contracts_server._dq_version_records
+load_records = contracts_server.load_records
+queue_flash = contracts_server.queue_flash
+save_records = contracts_server.save_records
+scenario_run_rows = contracts_server.scenario_run_rows
+store = contracts_server.store
+
+contracts_app = contracts_server.app
+demo_app = demo_server.app
 
 
 try:  # pragma: no cover - optional dependency in CI
@@ -23,14 +26,14 @@ except ModuleNotFoundError:  # pragma: no cover - fallback when pyspark missing
 
 
 def test_contracts_page():
-    client = TestClient(app)
+    client = TestClient(contracts_app)
     resp = client.get("/contracts")
     assert resp.status_code == 200
 
 
 def test_contract_detail_page():
     rec = load_records()[0]
-    client = TestClient(app)
+    client = TestClient(contracts_app)
     resp = client.get(f"/contracts/{rec.contract_id}/{rec.contract_version}")
     assert resp.status_code == 200
     assert 'id="access-tab"' in resp.text
@@ -39,7 +42,7 @@ def test_contract_detail_page():
 
 def test_contract_versions_page():
     rec = load_records()[0]
-    client = TestClient(app)
+    client = TestClient(contracts_app)
     resp = client.get(f"/contracts/{rec.contract_id}")
     assert resp.status_code == 200
     assert "Open editor" in resp.text
@@ -47,7 +50,7 @@ def test_contract_versions_page():
 
 
 def test_contract_edit_form_renders_editor_sections():
-    client = TestClient(app)
+    client = TestClient(contracts_app)
     resp = client.get("/contracts/orders/1.1.0/edit")
     assert resp.status_code == 200
     assert "Contract basics" in resp.text
@@ -58,7 +61,7 @@ def test_contract_edit_form_renders_editor_sections():
 
 
 def test_new_contract_form_defaults():
-    client = TestClient(app)
+    client = TestClient(contracts_app)
     resp = client.get("/contracts/new")
     assert resp.status_code == 200
     assert "Contract basics" in resp.text
@@ -68,13 +71,13 @@ def test_new_contract_form_defaults():
 
 
 def test_customers_contract_versions_page():
-    client = TestClient(app)
+    client = TestClient(contracts_app)
     resp = client.get("/contracts/customers")
     assert resp.status_code == 200
 
 
 def test_pipeline_runs_page_lists_scenarios():
-    client = TestClient(app)
+    client = TestClient(demo_app)
     resp = client.get("/pipeline-runs")
     assert resp.status_code == 200
     for key, cfg in SCENARIOS.items():
@@ -85,7 +88,7 @@ def test_pipeline_runs_page_lists_scenarios():
 
 def test_dataset_detail_page():
     rec = load_records()[0]
-    client = TestClient(app)
+    client = TestClient(contracts_app)
     resp = client.get(f"/datasets/{rec.dataset_name}/{rec.dataset_version}")
     assert resp.status_code == 200
     assert "order_id" in resp.text
@@ -93,13 +96,13 @@ def test_dataset_detail_page():
 
 def test_dataset_versions_page():
     rec = load_records()[0]
-    client = TestClient(app)
+    client = TestClient(contracts_app)
     resp = client.get(f"/datasets/{rec.dataset_name}")
     assert resp.status_code == 200
 
 
 def test_datasets_page_catalog_overview():
-    client = TestClient(app)
+    client = TestClient(contracts_app)
     resp = client.get("/datasets")
     assert resp.status_code == 200
     assert "orders" in resp.text
@@ -120,7 +123,7 @@ def test_dataset_pages_without_contract():
         violations=0,
     )
     save_records([*original, record])
-    client = TestClient(app)
+    client = TestClient(contracts_app)
     try:
         resp = client.get("/datasets")
         assert resp.status_code == 200
@@ -139,7 +142,7 @@ def test_dataset_pages_without_contract():
 
 def test_flash_message_consumed_once():
     token = queue_flash(message="Hello there", error=None)
-    client = TestClient(app)
+    client = TestClient(demo_app)
 
     first = client.get(f"/pipeline-runs?flash={token}")
     assert first.status_code == 200
@@ -282,7 +285,7 @@ def test_scenario_rows_ignore_mismatched_scenario_runs():
 @pytest.mark.skipif(not PYSPARK_AVAILABLE, reason="pyspark required for preview API")
 def test_contract_preview_api():
     rec = load_records()[0]
-    client = TestClient(app)
+    client = TestClient(contracts_app)
     resp = client.get(f"/api/contracts/{rec.contract_id}/{rec.contract_version}/preview")
     assert resp.status_code == 200
     payload = resp.json()
